% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcs.R
\name{BetweenGroup.test}
\alias{BetweenGroup.test}
\title{Perform between group statistical tests on Raman data}
\usage{
BetweenGroup.test(data, group, p.adj.method = "bonferroni", paired = FALSE)
}
\arguments{
\item{data}{A numeric matrix or data frame containing the Raman data.}

\item{group}{A factor indicating the grouping variable.}

\item{p.adj.method}{The method used for p-value adjustment. Default is "bonferroni".}

\item{paired}{Logical indicating whether paired tests should be performed for t-tests and Wilcoxon tests. Default is FALSE.}
}
\value{
A data frame with the results of the statistical tests.
}
\description{
This function performs between group statistical tests on Raman data,
including t-tests, Wilcoxon tests, variances tests, Oneway tests, and Kruskal-Wallis tests.
}
