% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality_control_jump.R
\name{bubblefill}
\alias{bubblefill}
\title{Fill bubbles in a spectrum to correct baseline}
\usage{
bubblefill(spectrum, min_bubble_widths = 50)
}
\arguments{
\item{spectrum}{A numeric vector representing the spectrum.}

\item{min_bubble_widths}{A numeric value or vector specifying the minimum width of the bubbles.
If a vector, it should be of the same length as the spectrum, specifying the minimum width for each point.}
}
\value{
A list containing the corrected spectrum (raman), the peak locations (peaks), and the band boundaries (bands).
}
\description{
This function applies a bubble-filling algorithm to a spectrum to correct the baseline.
It uses the `bubbleloop` function to identify and adjust the baseline, and then
identifies the peaks in the spectrum after baseline correction.
}
