% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamEx-Loaddata.R
\name{build_ramanome_object}
\alias{build_ramanome_object}
\title{Build a Ramanome Object from Parsed Spectral Data}
\usage{
build_ramanome_object(
  data_matrix,
  wavenumber,
  file_infor = NULL,
  meta.data = NULL,
  group.index = 1,
  group.names = "Group",
  group.levels = NULL,
  group_splits = "/|_"
)
}
\arguments{
\item{data_matrix}{A numeric matrix of Raman spectral intensities.
Rows correspond to spectra (samples), and columns correspond to wavenumbers.}

\item{wavenumber}{A numeric vector giving the wavenumber values associated with each spectral column.}

\item{file_infor}{Optional character vector of source file informations.
If provided, these are used to derive sample grouping and for naming rows.}

\item{meta.data}{Optional data frame of metadata.
If supplied, it will override automatically generated grouping information.}

\item{group.index}{Integer vector specifying which elements from filename splits to use as grouping variables.}

\item{group.names}{Character vector giving names to group columns.}

\item{group.levels}{Optional character vector of predefined factor levels.}

\item{group_splits}{Character pattern (regular expression) used for splitting filenames (default: `"/|_"`).}
}
\value{
A single `Ramanome` object.
}
\description{
Constructs a standardized **Ramanome** object from a spectral intensity matrix 
and its corresponding wavenumber vector.  

This function enables manually assemble of a `Ramanome` object 
from pre‑processed spectral data, meta.data.
}
\examples{
\dontrun{
# Build a Ramanome object from a supplied spectral matrix
raw_mat <- matrix(rnorm(10 * 2000), nrow = 10)   # 10 spectra × 2000 wavenumbers
wnum <- seq(500, 3150, length.out = 2000)
files <- paste0("Sample_", 1:10)
Raman <- build_ramanome_object(raw_mat, wnum, filenames = files)
}
}
