% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion-matrix.R
\name{cal_accuracy}
\alias{cal_accuracy}
\title{Calculate accuracy of prediction matrix}
\usage{
cal_accuracy(pred_matrix)
}
\arguments{
\item{pred_matrix}{The prediction matrix}
}
\value{
The accuracy matrix
}
\description{
This function calculates the accuracy of a prediction matrix.
}
