% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamEx-Class.R
\name{$,Ramanome-method}
\alias{$,Ramanome-method}
\title{Access slots of Ramanome object using $ operator}
\usage{
\S4method{$}{Ramanome}(x, name)
}
\arguments{
\item{x}{A Ramanome object}

\item{name}{The name of the slot to access}
}
\value{
The requested slot content
}
\description{
This method allows accessing slots of a Ramanome object using the $ operator.
It can access datasets, meta.data columns, reductions, and interested.bands.
}
