% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{Classification.Gmm}
\alias{Classification.Gmm}
\title{Gaussian Mixture Model (GMM)}
\usage{
Classification.Gmm(train, test = NULL)
}
\arguments{
\item{train}{The training data object}

\item{test}{The test data object (optional)}
}
\value{
The GMM model.
}
\description{
A probabilistic model that assumes data is generated from
a mixture of Gaussian distributions and assigns probabilities
 to each class
}
