% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{Classification.Rf}
\alias{Classification.Rf}
\title{Perform classification use RandomForest model}
\usage{
Classification.Rf(train, test = NULL, show = TRUE, save = FALSE, seed = 42)
}
\arguments{
\item{train}{The training data object}

\item{test}{The test data object (optional)}

\item{show}{Whether user want to show the results}

\item{save}{Wether user want to save the results}

\item{seed}{The random seed}
}
\description{
This function performs classification using RF
}
