% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{Classification.Svm}
\alias{Classification.Svm}
\title{Perform classification use svm model}
\usage{
Classification.Svm(train, test = NULL, show = TRUE, save = FALSE, seed = 42)
}
\arguments{
\item{train}{The training data object}

\item{test}{The test data object (optional)}

\item{show}{Whether user want to show the results}

\item{save}{Wether user want to save the results}

\item{seed}{The random seed}
}
\value{
A list containing:
  \item{pred.train}{The confusion matrix plot for training data}
  \item{pred.test}{The confusion matrix plot for test data}
}
\description{
This function performs classification using svm
}
\examples{
# Load example data
data_file <- system.file("extdata", "data", package = "RamEx")
RamEx_data <- read.spec.load(data_file, group.index = 2)

# Perform SVM classification with auto split
Classification.Svm(RamEx_data)
}
