% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_paired_markers.R
\name{compute_correlations}
\alias{compute_correlations}
\title{Compute Correlations with a Target Variable}
\usage{
compute_correlations(X, y, min.cor = 0.8)
}
\arguments{
\item{X}{A data frame containing the dataset.}

\item{y}{A vector containing the target variable.}

\item{min.cor}{The minimum correlation threshold. Defaults to 0.8.}
}
\value{
A list containing the high correlation elements, individual correlations,
correlations for two-variable combinations, and correlations for three-variable combinations.
}
\description{
This function calculates the correlations between each column of a dataset X and a target
variable y. It also computes the correlations for combinations of two and three
variables, after applying certain transformations.
}
