% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_value.R
\name{confirm.name}
\alias{confirm.name}
\title{Confirm the selected wavenumber or Raman band}
\usage{
confirm.name(waves)
}
\arguments{
\item{waves}{A vector with one or two values representing the wavenumber or wavenumber range.}
}
\value{
The name of the selected wavenumber or Raman band.
}
\description{
This function confirms the selected wavenumber or Raman band and returns the corresponding name.
}
