% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_value.R
\name{confirm.select}
\alias{confirm.select}
\title{Confirm the selected wavenumber or Raman band}
\usage{
confirm.select(object, waves)
}
\arguments{
\item{object}{A Ramanome object.}

\item{waves}{The selected waves.}
}
\description{
This function confirms the selected wavenumber or Raman band and returns the corresponding name.
}
