% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-spike.r
\name{convolve_matrix}
\alias{convolve_matrix}
\title{find spikes}
\usage{
convolve_matrix(mat, kernel)
}
\arguments{
\item{mat}{= matrix}
}
\value{
a spike_list(sample_i_ind and wavenumber_ind)
}
\description{
find spikes
}
\examples{
pre_spike_find(hyperspec, width = 5)

}
