% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality_control_Other_methods.R
\name{covFastMCD}
\alias{covFastMCD}
\title{Fast Minimum Covariance Determinant Estimation}
\usage{
covFastMCD(x, alpha, m, l, delta)
}
\arguments{
\item{x}{The input dataset, a matrix or data frame.}

\item{alpha}{The significance level for the MCD estimation.}

\item{m}{The number of initial subsets to consider.}

\item{l}{The number of final subsets to consider.}

\item{delta}{The quantile for the reweighting step.}
}
\value{
A list containing the raw center, raw covariance matrix, best subset indices,
weights for reweighting, and the reweighted center and covariance matrix.
}
\description{
This function implements the Fast Minimum Covariance Determinant (MCD) method to estimate
the covariance matrix and identify potential outliers in the data. The MCD method is
robust to outliers and can be used for multivariate data.
}
