% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-spec.R
\name{cut.spec}
\alias{cut.spec}
\title{Cut spectral data based on cutoff values}
\usage{
\method{cut}{spec}(single_spec, cutoff)
}
\arguments{
\item{single_spec}{A data frame containing the spectral data.}

\item{cutoff}{A numeric vector with two values representing the lower and upper bound cutoff values.}
}
\value{
The subset of spectral data within the specified cutoff range.
}
\description{
This function cuts the spectral data based on cutoff values for the wavenumber range.
}
