% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-new-class.R
\docType{class}
\name{DataMatrix-class}
\alias{DataMatrix-class}
\title{DataMatrix class}
\description{
This class represents the DataMatrix.
}
\section{Slots}{

\describe{
\item{\code{wavenumber}}{A numeric vector of wavenumbers}

\item{\code{label}}{A factor representing the labels}

\item{\code{filenames}}{Character vector of filenames}

\item{\code{spec}}{A matrix of spectra data}
}}

