% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality_control_Other_methods.R
\name{detect_T2}
\alias{detect_T2}
\title{Detect outliers in a dataset using Hotelling's T2 method}
\usage{
detect_T2(x)
}
\arguments{
\item{x}{A matrix of spectral data where each row represents a single spectrum.}
}
\value{
A data frame containing two columns: 'out' indicating whether each spectrum is an outlier,
and 'dis' containing the corresponding Hotelling's T2 distance values.
}
\description{
This function identifies outliers in a dataset by applying Hotelling's T2 statistical method.
It uses parallel processing to speed up the computation.
}
