% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Package Initialization Function}
\usage{
.onLoad(libname, RamEx)
}
\arguments{
\item{libname}{The library name}

\item{RamEx}{The package name}
}
\value{
None
}
\description{
This function is called when the package is loaded. It loads the required data into the global environment.
}
\keyword{internal}
