% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-new-class.R, R/draw-mean-spec.R
\name{draw.mean}
\alias{draw.mean}
\title{Draw mean spectrum}
\usage{
draw.mean(object, gap = NA)

draw.mean(object, gap = NA)
}
\arguments{
\item{object}{The Ramanome object}

\item{gap}{The gap between the mean spectra (default: 0)}
}
\value{
The plot of the mean spectrum

The plot of the mean spectrum
}
\description{
This function generates a plot of the mean spectrum for a Ramanome object.

This function generates a plot of the mean spectrum for a Ramanome object.
}
