% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature-Reduction.R
\name{Feature.Reduction.Tsne}
\alias{Feature.Reduction.Tsne}
\title{Perform t-SNE Dimensionality Reduction and Plotting}
\usage{
Feature.Reduction.Tsne(object, draw = TRUE, save = FALSE)
}
\arguments{
\item{object}{A Ramanome object containing the dataset and metadata.}

\item{draw}{A logical value indicating whether to draw the plot. Defaults to TRUE.}

\item{save}{A logical value indicating whether to save the plot as a PNG file.
Defaults to FALSE.}
}
\value{
The updated Ramanome object with the t-SNE results appended to the reductions slot.
}
\description{
This function performs t-SNE (t-Distributed Stochastic Neighbor Embedding) on a
dataset extracted from a Ramanome object. It reduces the dimensionality to two
dimensions and creates a plot of the results. The plot can be customized with various
themes and labels.
}
\examples{
data(RamEx_data)
data_smoothed <- Preprocessing.Smooth.Sg(RamEx_data)
data_baseline <- Preprocessing.Baseline.Polyfit(data_smoothed)
data_baseline_bubble <- Preprocessing.Baseline.Bubble(data_smoothed)
data_normalized <- Preprocessing.Normalize(data_baseline, "ch")
data_cleaned <- Qualitycontrol.ICOD(data_normalized@datasets$normalized.data,var_tol = 0.4)
data_cleaned <- RamEx_data[qc_icod$quality,]
data.reduction.tsne <- Feature.Reduction.Tsne(data_cleaned, draw=TRUE, save = FALSE)
}
