% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature-Reduction.R
\name{Feature.Reduction.Umap}
\alias{Feature.Reduction.Umap}
\title{Perform UMAP Dimensionality Reduction}
\usage{
Feature.Reduction.Umap(
  object,
  PCA = 20,
  n_pc = 2,
  n_neighbors = 30,
  min.dist = 0.01,
  spread = 1,
  show = TRUE,
  save = FALSE,
  seed = 42
)
}
\arguments{
\item{object}{A Ramanome object containing the dataset.}

\item{PCA}{A numeric value indicating the number of principal components to use. Defaults to 20 (if NULL,raw data matrix will be used to generate UMAP plot).}

\item{n_pc}{The number of UMAP components to save. Defaults to 2.}

\item{n_neighbors}{Controls the balance between local and global structure in UMAP. Higher values capture more global structure, while lower values focus on local details (typical range: 5–50).}

\item{min.dist}{Determines the minimum distance between points in the embedding. Lower values (e.g., 0.1) produce tighter clusters, while higher values (e.g., 0.5) allow more spread.}

\item{spread}{Scales the effective scale of points in the embedding. Works with min.dist to control cluster density—higher values spread points apart, while lower values compress them.}

\item{show}{A logical value indicating whether to draw the plot. Defaults to TRUE.}

\item{save}{A logical value indicating whether to save the plot as a PNG file. Defaults to FALSE.}

\item{seed}{A numeric value indicating the seed for the random number generator. Defaults to 42.}
}
\value{
The updated Ramanome object with the UMAP reduction results.
}
\description{
This function performs Uniform Manifold Approximation and Projection (UMAP) on the
dataset stored in a Ramanome object. It first applies PCA to reduce the dataset
to 20 dimensions and then applies UMAP to reduce it to two dimensions. The reduced
data is stored in the object and a plot is generated if specified.
}
\examples{
data(RamEx_data)
data_processed <- Preprocessing.OneStep(RamEx_data)
data.reduction.umap <- Feature.Reduction.Umap(data_processed, show=TRUE, save = FALSE)

}
