% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_paired_markers.R
\name{find_markers_roc}
\alias{find_markers_roc}
\title{Find Markers Using ROC Analysis}
\usage{
find_markers_roc(matrix, group, threshold = 0.75)
}
\arguments{
\item{matrix}{A matrix of spectral data where each row represents a sample and each
column represents a wavelength.}

\item{group}{A vector of group identifiers corresponding to each row in the matrix.}

\item{threshold}{The minimum AUC threshold for considering a marker significant.
Defaults to 0.75.}
}
\value{
A list containing two data frames: 'markers_single' for single markers and
'markers_paired' for paired markers, each with their corresponding AUC scores.
}
\description{
This function performs Receiver Operating Characteristic (ROC) analysis to identify
markers in a dataset that are associated with different groups. It calculates the
Area Under the Curve (AUC) for single markers and paired markers.
}
