% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raman-Markers.R
\name{get.mislabel.scores}
\alias{get.mislabel.scores}
\title{Calculate mislabel scores for classification results}
\usage{
get.mislabel.scores(y, y.prob)
}
\arguments{
\item{y}{A vector containing the true labels.}

\item{y.prob}{A matrix containing class probabilities.}
}
\value{
A matrix containing mislabel scores for each class.
}
\description{
This function calculates mislabel scores based on class probabilities for classification results.
}
\examples{
# Create sample data
y <- factor(c("A", "A", "B", "B", "A"))
y.prob <- matrix(c(0.8, 0.7, 0.3, 0.2, 0.6,
                   0.2, 0.3, 0.7, 0.8, 0.4),
                 nrow=5, ncol=2,
                 dimnames=list(NULL, c("A", "B")))

}
