% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raman-Markers.R
\name{get.oob.probability.from.forest}
\alias{get.oob.probability.from.forest}
\title{Get out-of-bag class probabilities from a random forest model}
\usage{
get.oob.probability.from.forest(model, x)
}
\arguments{
\item{model}{The random forest model object.}

\item{x}{A matrix or data frame containing the predictors.}
}
\value{
A matrix with the predicted class probabilities for each sample.
}
\description{
This function calculates the out-of-bag class probabilities for each sample in the dataset, based on the aggregated class votes from the out-of-bag trees.
}
\examples{
# Create sample data and fit random forest model
set.seed(123)
x <- matrix(rnorm(200), nrow=20)
y <- factor(rep(c("A", "B"), each=10))
rf_model <- randomForest::randomForest(x, y, keep.inbag=TRUE)

}
