% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raman-Markers.R
\name{get.oob.votes.from.forest}
\alias{get.oob.votes.from.forest}
\title{Get Out-of-Bag class votes from a random forest model}
\usage{
get.oob.votes.from.forest(model, x)
}
\arguments{
\item{model}{The random forest model.}

\item{x}{A matrix or data frame containing the predictor variables.}
}
\value{
A matrix of class votes for each sample, with rows representing samples and columns representing classes.


# Get the Out-of-Bag class votes for a new data set
x_new <- iris[-1]
votes <- get.oob.votes.from.forest(model, x_new)
}
\description{
This function extracts the aggregated class votes for each sample using only the Out-of-Bag trees from a random forest model.
}
