% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocesssing.R
\name{gpupre_spike_matrix}
\alias{gpupre_spike_matrix}
\title{Pre spike the spec. with Rcpp}
\usage{
gpupre_spike_matrix(all_data, device = NULL)
}
\arguments{
\item{all_data}{The rammannome class}

\item{device}{Parallel device, default is null for CPU, can be set to "GPU"}
}
\value{
A rammanome object
}
\description{
This function performs Pre spike function
}
\examples{
# Create sample data
data <- matrix(rnorm(1000), nrow=10)

}
