% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality-Control.R
\name{grow_bubble}
\alias{grow_bubble}
\title{Grow a Bubble Chart}
\usage{
grow_bubble(spectrum, alignment = "center")
}
\arguments{
\item{spectrum}{A numeric vector representing the spectrum.}

\item{alignment}{A character string specifying the alignment of the bubble, which can be "left", "right", or "center".}
}
\value{
A list containing the bubble values and the relative touching point.
}
\description{
This function generates a bubble chart representation of a spectrum by expanding a bubble
until it touches the spectrum. The bubble can be aligned to the left, right, or centered.
}
