% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering-Louvain-leiden.R
\name{hca_clustering}
\alias{hca_clustering}
\title{Perform Hierarchical Clustering Analysis (HCA)}
\usage{
hca_clustering(object)
}
\arguments{
\item{object}{A Ramanome object containing the dataset.}
}
\value{
The hierarchical clustering fit object.
}
\description{
This function performs hierarchical clustering on the dataset retrieved from a
Ramanome object using the average linkage method. It calculates the Euclidean distance
matrix and then applies the hierarchical clustering algorithm to create a dendrogram.
The function also plots the resulting dendrogram.
}
