% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamEx-Basic-Funciton.R
\name{image.peak}
\alias{image.peak}
\title{Generate Raman imaging}
\usage{
\method{image}{peak}(object, peak)
}
\arguments{
\item{object}{The Ramanome object}

\item{peak}{The specified peak for Raman imaging}
}
\value{
NULL. The function saves a JPEG file named 'Raman_Imaging_[peak].jpeg' in the current working directory.
        The image shows a heatmap-style visualization of the Raman intensity distribution at the specified peak.
}
\description{
This function generates a Raman imaging plot for a specified peak in a Ramanome object.
}
