% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_value.R
\name{intensity}
\alias{intensity}
\title{Calculate Raman intensities for selected wavenumbers or Raman bands}
\usage{
intensity(object, wavenumber)
}
\arguments{
\item{object}{A Ramanome object.}

\item{wavenumber}{A list of wavenumbers or wavenumber ranges.}
}
\value{
The updated Ramanome object.
}
\description{
This function calculates the Raman intensities for the selected wavenumbers or Raman bands and updates the Ramanome object.
}
\examples{
data(Ramanome)
raman <- intensity(Ramanome, wavenumber = list(600, c(800, 1000)))
}
