% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intra-Ramanome-Analysis.R
\name{Intraramanome.Analysis.2Dcos}
\alias{Intraramanome.Analysis.2Dcos}
\title{2D Correlation Spectroscopy Analysis (2D-COS)
Captures both synchronous (simultaneous changes) and asynchronous (sequential changes) relationships, providing detailed insights into spectral changes}
\usage{
Intraramanome.Analysis.2Dcos(object)
}
\arguments{
\item{object}{A Ramanome object.}
}
\value{
synchronous and asynchronous correlation spectra result
}
\description{
2D Correlation Spectroscopy Analysis (2D-COS)
Captures both synchronous (simultaneous changes) and asynchronous (sequential changes) relationships, providing detailed insights into spectral changes
}
\examples{
data(RamEx_data)
data_processed <- Preprocessing.OneStep(RamEx_data)
corr_2d <- Intraramanome.Analysis.2Dcos(data_processed)
}
