% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intra-Ramanome-Analysis.R
\name{Intraramanome.Analysis.Irca.Global.draw}
\alias{Intraramanome.Analysis.Irca.Global.draw}
\title{Perform IRCA analysis on global scale}
\usage{
Intraramanome.Analysis.Irca.Global.draw(dataset, group)
}
\arguments{
\item{dataset}{The input dataset for analysis}

\item{group}{The group information for the dataset}
}
\value{
A dataframe containing the significant correlations between variables
}
\description{
This function performs IRCA analysis on the global scale using the input dataset and group information.
It generates a chord diagram that represents the significant correlations between variables.
}
\examples{
# Create sample dataset
n <- 100  # samples
p <- 20   # variables
dataset <- matrix(rnorm(n*p), nrow=n)
colnames(dataset) <- paste0("V", 1:p)
group <- rep(c("A", "B"), each=n/2)

}
