% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intra-Ramanome-Analysis.R
\name{Intraramanome.Analysis.Irca.Global}
\alias{Intraramanome.Analysis.Irca.Global}
\title{Perform global IRCA analysis}
\usage{
Intraramanome.Analysis.Irca.Global(object)
}
\arguments{
\item{object}{The Raman spectroscopy data object.}
}
\value{
A matrix containing the global IRCA interests.
}
\description{
This function performs global IRCA analysis on the given Raman spectroscopy data object.
It generates negative IRCA plots for each group and saves them as jpeg files.
}
