% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irca.R
\name{irca.global.draw}
\alias{irca.global.draw}
\title{Perform IRCA analysis on global scale}
\usage{
irca.global.draw(dataset, group)
}
\arguments{
\item{dataset}{The input dataset for analysis}

\item{group}{The group information for the dataset}
}
\value{
A dataframe containing the significant correlations between variables
}
\description{
This function performs IRCA analysis on the global scale using the input dataset and group information.
It generates a chord diagram that represents the significant correlations between variables.
}
