% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irca.R
\name{irca.local.draw}
\alias{irca.local.draw}
\title{Perform IRCA analysis on local scale}
\usage{
irca.local.draw(dataset, bands_ann)
}
\arguments{
\item{dataset}{The input dataset for analysis}

\item{bands_ann}{The bands annotation information}
}
\description{
This function performs IRCA analysis on the local scale using the input dataset and bands annotation information.
It generates a chord diagram that represents the significant correlations between variables within each band.
}
