% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irca.R
\name{irca.local}
\alias{irca.local}
\title{Perform local IRCA analysis}
\usage{
irca.local(object, bands_ann)
}
\arguments{
\item{object}{The Raman spectroscopy data object.}

\item{bands_ann}{The band annotation data.}
}
\description{
This function performs local IRCA analysis on the given Raman spectroscopy data object.
It generates negative IRCA plots for each group and saves them as png files.
}
