% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality-Control.R
\name{keep_largest}
\alias{keep_largest}
\title{Keep the largest values between baseline and bubble}
\usage{
keep_largest(baseline, bubble)
}
\arguments{
\item{baseline}{A vector of baseline values.}

\item{bubble}{A vector of bubble values.}
}
\value{
A vector of the larger values
}
\description{
This function compares the baseline and bubble values and keeps the larger ones.
}
