% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intra-Ramanome-Analysis.R
\name{mahalDistForEvery}
\alias{mahalDistForEvery}
\title{Calculate Mahalanobis distance for every observation}
\usage{
mahalDistForEvery(data)
}
\arguments{
\item{data}{The input data matrix.}
}
\value{
A vector containing the Mahalanobis distance for each observation.
}
\description{
This function calculates the Mahalanobis distance for every observation in the given data matrix.
}
\examples{
# Create a sample data matrix
data <- matrix(rnorm(100), nrow=20)
}
