% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCR-ALS.R
\name{MCRALS}
\alias{MCRALS}
\title{Perform Multivariate Curve Resolution Alternating Least Squares (MCR-ALS)}
\usage{
MCRALS(object, ncomp)
}
\arguments{
\item{object}{A Ramanome object containing the dataset.}

\item{ncomp}{The number of components to estimate.}
}
\value{
A list containing the estimated components and concentrations.
}
\description{
This function applies the MCR-ALS algorithm to the dataset stored in a Ramanome object.
MCR-ALS is a chemometric technique used to resolve overlapping spectra into their
constituent pure components. The function returns the estimated components and
concentrations.
}
