% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-new-class.R, R/draw-mean-spec.R
\name{mean.spec}
\alias{mean.spec}
\title{Generate mean spectrum}
\usage{
mean.spec(data, group, gap = 0.3)

mean.spec(data, group, gap = 0.3)
}
\arguments{
\item{data}{The data matrix}

\item{group}{The group factor}

\item{gap}{The gap between the mean spectra (default: 0.3)}
}
\value{
The mean spectrum

The mean spectrum
}
\description{
This function calculates the mean spectrum for a given Ramanome object.

This function calculates the mean spectrum for a given Ramanome object.
}
