% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamEx-Class.R
\name{names,Ramanome-method}
\alias{names,Ramanome-method}
\title{Get available slot names for Ramanome object}
\usage{
\S4method{names}{Ramanome}(x)
}
\arguments{
\item{x}{A Ramanome object}
}
\value{
A character vector of available slot names
}
\description{
This method returns all available slot names that can be accessed using the $ operator.
It enables tab-completion in RStudio and other IDEs.
}
