% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality-Control.R
\name{outliers_mcdEst}
\alias{outliers_mcdEst}
\title{Estimate outliers using the Minimum Covariance Determinant (MCD) method}
\usage{
outliers_mcdEst(x, index_good, h = 0.75, alpha = 0.01, na.rm = TRUE)
}
\arguments{
\item{x}{The input dataset.}

\item{index_good}{A vector indicating the indices of good observations to use for the estimation.}

\item{h}{The fraction of the data to use for computing the MCD (between 0 and 1). Defaults to 0.75.}

\item{alpha}{The significance level for the MCD. Defaults to 0.01.}

\item{na.rm}{A logical value indicating whether to remove NA values. Defaults to TRUE.}
}
\value{
A list containing the following components:
  - MaxDist: The maximum distance threshold for outliers.
  - center: The estimated centroid of the data.
  - outliers_pos: The positions of the outliers.
  - outliers_val: The values of the outliers.
  - dist_from_center: The distances of each observation from the centroid.
}
\description{
This function implements the Minimum Covariance Determinant method to estimate
outliers in a dataset. It calculates the MCD estimate of scatter and location,
and identifies observations that are likely to be outliers based on their distance
from the centroid.
}
\examples{
# Create sample data with outliers
set.seed(123)
n <- 100
p <- 3

# Generate normal data
good_data <- matrix(rnorm(n*p), ncol=p)

# Add some outliers
outliers <- matrix(rnorm(10*p, mean=5), ncol=p)
data <- rbind(good_data, outliers)

# Mark initial good observations (all TRUE for initial estimation)
index_good <- rep(TRUE, nrow(data))

}
