% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotype-Analysis.R
\name{Phenotype.Analysis.Gmm}
\alias{Phenotype.Analysis.Gmm}
\title{Gaussian Mixture Model (GMM)}
\usage{
Phenotype.Analysis.Gmm(object, n_pc = 20, seed = 42)
}
\arguments{
\item{object}{The Ramanome object}

\item{n_pc}{The number of principal components to use}
}
\value{
A vector containing the cluster assignments for each data point
}
\description{
A probabilistic model that assumes data is generated from a mixture of Gaussian distributions. 
Models are estimated by EM algorithm initialized by hierarchical model-based agglomerative clustering. The optimal model is then selected according to BIC.
}
\examples{
data(RamEx_data)
data_processed <- Preprocessing.OneStep(RamEx_data)
cluster_gmm <- Phenotype.Analysis.Gmm(data_processed)
}
