% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotype-Analysis.R
\name{Phenotype.Analysis.Hca}
\alias{Phenotype.Analysis.Hca}
\title{Perform Hierarchical Clustering Analysis (HCA)}
\usage{
Phenotype.Analysis.Hca(object)
}
\arguments{
\item{object}{A Ramanome object containing the dataset.}
}
\value{
The hierarchical clustering fit object.
}
\description{
This function performs hierarchical clustering on the dataset retrieved from a
Ramanome object using the average linkage method. It calculates the Euclidean distance
matrix and then applies the hierarchical clustering algorithm to create a dendrogram.
The function also plots the resulting dendrogram.
}
\examples{
data(RamEx_data)
data_smoothed <- Preprocessing.Smooth.Sg(RamEx_data)
data_baseline <- Preprocessing.Baseline.Polyfit(data_smoothed)
data_baseline_bubble <- Preprocessing.Baseline.Bubble(data_smoothed)
data_normalized <- Preprocessing.Normalize(data_baseline, "ch")
qc_icod <- Qualitycontrol.ICOD(data_normalized@datasets$normalized.data,var_tol = 0.4)
data_cleaned <- data_normalized[qc_icod$quality,]
data_cleaned <- Feature.Reduction.Intensity(data_cleaned, list(c(2000,2250),c(2750,3050), 1450, 1665))
clusters_hca <- Phenotype.Analysis.Hca(data_cleaned)
}
