% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotype-Analysis.R
\name{Phenotype.Analysis.Leiden}
\alias{Phenotype.Analysis.Leiden}
\title{Leiden Clustering}
\usage{
Phenotype.Analysis.Leiden(
  object,
  resolutions,
  n_pc = 10,
  threshold = 0.001,
  k = 30,
  n_tree = 50,
  prune.SNN = 1/15,
  seed = 42,
  objective_function = c("modularity", "CPM"),
  n_iterations = 2L,
  verbose = TRUE
)
}
\description{
Build a shared nearest neighbor (SNN) graph from the latest spectral matrix
stored in a Ramanome object and run Louvain community detection across a
vector of resolution values.
}
