% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-ramanome-class.R
\name{plot,Ramanome,character-method}
\alias{plot,Ramanome,character-method}
\title{Plot method for Ramanome objects}
\usage{
\S4method{plot}{Ramanome,character}(x, y, ...)
}
\arguments{
\item{x}{The Ramanome object}

\item{y}{Character specifying the plot type}
}
\description{
Plot method for Ramanome objects
}
