% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.r
\name{Plot.Cluster.Distribution}
\alias{Plot.Cluster.Distribution}
\title{Plot histogram of group distributions}
\usage{
Plot.Cluster.Distribution(
  group_1,
  group_2,
  cols = RamEx.colors,
  width = 0.9,
  alpha = 0.5,
  curve_type = "sine"
)
}
\arguments{
\item{group_1}{A vector of group assignments}

\item{group_2}{A vector of group assignments}

\item{cols}{Optional color palette for groups}

\item{width}{Width of the alluvium (default: 0.9)}

\item{alpha}{Transparency of the alluvium (default: 0.5)}

\item{curve_type}{Type of curve for the alluvium (default: "sine")}
}
\value{
A ggplot object
}
\description{
Plot histogram of group distributions
}
