% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.r
\name{Plot.Heatmap.Markers}
\alias{Plot.Heatmap.Markers}
\title{Plot Heatmap of Markers}
\usage{
Plot.Heatmap.Markers(
  object,
  markers,
  group = object$group,
  scale = "row",
  show_legend = TRUE,
  cluster_cols = FALSE,
  cluster_rows = TRUE,
  show_rownames = FALSE,
  show_colnames = TRUE,
  show = TRUE,
  save = FALSE,
  wave_gap_threshold = 20,
  width = 10,
  height = 10
)
}
\arguments{
\item{object}{A Ramanome object}

\item{markers}{A list of markers (wave numbers)}

\item{group}{group information to show the heatmap and marker intensities}

\item{scale}{Whether to scale the data (default: "row")}

\item{show_legend}{Whether to show legend (default: TRUE)}

\item{show_rownames}{Whether to show row names (Wave Number) (default: FALSE)}

\item{show_colnames}{Whether to show column names (Group levels) (default: TRUE)}

\item{wave_gap_threshold}{Threshold for wave number gap to split wave number into groups (default: 50)}
}
\value{
A heatmap plot
}
\description{
Plot Heatmap of Markers
}
