% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.r
\name{Plot.Markers.Spectrum}
\alias{Plot.Markers.Spectrum}
\title{Mark the Ramanome Markers on Average Spectrum
Mark the Ramanome Markers on Average Spectrum}
\usage{
Plot.Markers.Spectrum(
  object,
  markers,
  gap = 0.6,
  cols = c("#F18656", "#397FC7")
)
}
\arguments{
\item{object}{A Ramanome object}

\item{gap}{Vertical gap between spectra (default: 0.6)}

\item{cols}{Color palette for markers (default: c('#F18656','#397FC7'))}

\item{markers_single}{A list of single markers (wave numbers)}

\item{markers_paired}{A list of paired markers (wave numbers) or NULL/NA}
}
\value{
A ggplot object
}
\description{
Mark the Ramanome Markers on Average Spectrum
Mark the Ramanome Markers on Average Spectrum
}
