% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.r
\name{Plot.reductions}
\alias{Plot.reductions}
\title{Plot Scatter Plot from Reductions}
\usage{
Plot.reductions(
  object,
  reduction = "UMAP",
  dims = c(1, 2),
  group = object$group,
  cols = RamEx.colors,
  point_size = 1,
  point_alpha = 0.5
)
}
\arguments{
\item{object}{A Ramanome object}

\item{reduction}{Name of the reduction to plot (default: "UMAP")}

\item{cols}{Optional color palette for groups}

\item{point_size}{Size of points (default: 1)}

\item{point_alpha}{Transparency of points (default: 0.5)}
}
\value{
A ggplot object
}
\description{
Plot Scatter Plot from Reductions
}
