% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.r
\name{Plot.scatter}
\alias{Plot.scatter}
\title{Plot scatter plot}
\usage{
Plot.scatter(x, y, cols = NULL, point_size = 1, point_alpha = 0.5)
}
\arguments{
\item{x}{A vector of x values}

\item{y}{A vector of y values}

\item{cols}{Optional color palette for points}

\item{point_size}{Size of points (default: 1)}

\item{point_alpha}{Transparency of points (default: 0.5)}
}
\value{
A ggplot object
}
\description{
Plot scatter plot
}
