% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.r
\name{Plot.ViolinBox}
\alias{Plot.ViolinBox}
\title{Plot Violin and Box Plots}
\usage{
Plot.ViolinBox(data, group, cols = NULL, violin_width = 0.75, box_width = 0.12)
}
\arguments{
\item{data}{A data frame or matrix of data}

\item{group}{A vector or data frame of group information}

\item{cols}{Optional color palette for groups}

\item{violin_width}{Width of violin plot (default: 0.75)}

\item{box_width}{Width of box plot (default: 0.12)}
}
\value{
A ggplot object
}
\description{
Plot Violin and Box Plots
}
