% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irca.R
\name{Plot_global_chordDiagram_rpvalue}
\alias{Plot_global_chordDiagram_rpvalue}
\title{Plot Global Chord Diagram with Statistical Significance}
\usage{
Plot_global_chordDiagram_rpvalue(
  Corr_mat,
  a_degree = NULL,
  a_SumCorr = NULL,
  a_MeanCorr = NULL,
  main_title = NULL,
  Pos_Edge = FALSE,
  Neg_Edge = FALSE,
  Threshold = 0.6,
  bands_ann = NULL
)
}
\arguments{
\item{Corr_mat}{The correlation matrix for plotting the chord diagram.}

\item{a_degree}{A dataframe containing the degree information.}

\item{a_SumCorr}{A dataframe containing the sum correlation information.}

\item{a_MeanCorr}{A dataframe containing the mean correlation information.}

\item{main_title}{The main title for the plot.}

\item{Pos_Edge}{Logical indicating whether to include positive edges.}

\item{Neg_Edge}{Logical indicating whether to include negative edges.}

\item{Threshold}{The threshold value for considering edges as statistically significant.}

\item{bands_ann}{A dataframe containing the band annotation information.}
}
\value{
No explicit return value. The function generates the plot directly.
}
\description{
This function plots a global chord diagram based on a correlation matrix, highlighting edges with statistical significance.
}
