% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-baseline.R
\name{pre.baseline.bubble}
\alias{pre.baseline.bubble}
\title{Remove baseline by bubble}
\usage{
pre.baseline.bubble(object)
}
\arguments{
\item{object}{Ramanome object}
}
\value{
The baseline corrected spectra with bubble method
}
\description{
This function performs baseline pre-processing on the input object.
}
