% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-baseline.R
\name{pre.baseline}
\alias{pre.baseline}
\title{Preprocess baseline}
\usage{
pre.baseline(object, order = 1)
}
\arguments{
\item{object}{The input object}

\item{order}{The order of the polynomial used in the baseline calculation (default: 1)}
}
\value{
The modified object with preprocessed baseline data
}
\description{
This function performs baseline pre-processing on the input object.
}
