% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-normalize.R
\name{pre.normalize.area}
\alias{pre.normalize.area}
\title{Pre-process Raman data by area normalization method}
\usage{
pre.normalize.area(object)
}
\arguments{
\item{object}{A Ramanome object.}
}
\value{
The updated Ramanome object with the area normalized Raman data.
}
\description{
This function applies normalization to the Raman data based on the area method and updates the Raman object.
}
