% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-normalize.R
\name{pre.normalize.ch}
\alias{pre.normalize.ch}
\title{Pre-process Raman data by ch normalization method}
\usage{
pre.normalize.ch(object)
}
\arguments{
\item{object}{A Ramanome object.}
}
\value{
The updated Ramanome object with the ch normalized Raman data.
}
\description{
This function applies normalization to the Raman data based on the ch method and updates the Raman object.
}
