% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-normalize.R
\name{pre.normalize}
\alias{pre.normalize}
\title{Pre-process Raman data by normalization}
\usage{
pre.normalize(object, method, wave = NULL)
}
\arguments{
\item{object}{A Ramanome object.}

\item{method}{The normalization method to use: "CH", "max", "specific", or "area".}

\item{wave}{The specific wavenumber to use for the "specific" normalization method.}
}
\value{
The updated Ramanome object with the normalized Raman data.
}
\description{
This function applies normalization to the Raman data based on the specified method and updates the Raman object.
}
