% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-normalize.R
\name{pre.normalize.specific}
\alias{pre.normalize.specific}
\title{Pre-process Raman data by specific normalization}
\usage{
pre.normalize.specific(object, wave)
}
\arguments{
\item{object}{A Ramanome object.}

\item{wave}{The specific wavenumber to use for the "specific" normalization method.}
}
\value{
The updated Ramanome object with the specific normalized Raman data.
}
\description{
This function applies normalization to the Raman data based on the specific method and updates the Raman object.
}
