% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-smooth.R
\name{pre.smooth}
\alias{pre.smooth}
\title{Pre-process Raman data by smoothing with Savitzky-Golay filtering}
\usage{
pre.smooth(object, m = 0, p = 5, w = 11, delta.wav = 2)
}
\arguments{
\item{object}{A Ramanome object.}

\item{m}{The order of the polynomial to fit. Default is 0.}

\item{p}{The degree of the polynomial to fit. Default is 5.}

\item{w}{The width of the smoothing window. Default is 11.}

\item{delta.wav}{The wavelength difference between adjacent data points. Default is 2.}
}
\value{
The updated Ramanome object with the smoothed Raman data.
}
\description{
This function applies Savitzky-Golay filtering to smooth the Raman data and updates the Raman object.
}
