% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-smooth.R
\name{pre.smooth.snv}
\alias{pre.smooth.snv}
\title{Pre-process Raman data by smoothing with standardNormalVariate}
\usage{
pre.smooth.snv(object)
}
\arguments{
\item{object}{A Ramanome object.}
}
\description{
This function applies standardNormalVariate filtering to smooth the Raman data and updates the Raman object.
}
