% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-spike.r
\name{pre.spike}
\alias{pre.spike}
\title{Pre spike the spec. with Rcpp}
\usage{
pre.spike(object)
}
\arguments{
\item{object}{The rammanome class}
}
\value{
A rammanome object
}
\description{
This function performs Pre spike function
}
