% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantification.R
\name{predict_quantification}
\alias{predict_quantification}
\title{Predict using a saved quantification model}
\usage{
predict_quantification(model, new_data)
}
\arguments{
\item{model}{The saved quantification model (from Quantification.Pls, Quantification.Mlr, or Quantification.Glm)}

\item{new_data}{The new data object to predict}

\item{show}{Whether to show the plot of the results}

\item{save}{Whether to save the plot of the results (default path: getwd())}
}
\value{
A list containing:
\describe{
  \item{predictions}{The predicted values for the new data}
}
}
\description{
This function uses a saved quantification model to predict values for new data.
}
\examples{
data(RamEx_data)
# Train a model
model <- Quantification.Pls(RamEx_data)
# Use the model to predict new data
predictions <- predict_quantification(model, RamEx_data)
}
