% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocesssing.R
\name{Preprocessing.Baseline.Bubble}
\alias{Preprocessing.Baseline.Bubble}
\title{Remove baseline by bubble}
\usage{
Preprocessing.Baseline.Bubble(object)
}
\arguments{
\item{object}{Ramanome object}
}
\value{
The baseline corrected spectra with bubble method
}
\description{
This function performs baseline pre-processing on the input object.
}
\examples{
data(RamEx_data)
data_smoothed <- Preprocessing.Smooth.Sg(RamEx_data)
data_baseline <- Preprocessing.Baseline.Polyfit(data_smoothed)
data_baseline_bubble <- Preprocessing.Baseline.Bubble(data_smoothed)
}
