% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{Preprocessing.Baseline.Polyfit}
\alias{Preprocessing.Baseline.Polyfit}
\title{Preprocess baseline}
\usage{
Preprocessing.Baseline.Polyfit(object, order = 1)
}
\arguments{
\item{object}{The input object}

\item{order}{The order of the polynomial used in the baseline calculation (default: 1)}
}
\value{
The modified object with preprocessed baseline data
}
\description{
This function performs baseline pre-processing on the input object.
}
\examples{
data(RamEx_data)
data_smoothed <- Preprocessing.Smooth.Sg(RamEx_data)
data_baseline <- Preprocessing.Baseline.Polyfit(data_smoothed)
}
