% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{Preprocessing.Cutoff}
\alias{Preprocessing.Cutoff}
\title{Spectral truncation}
\usage{
Preprocessing.Cutoff(object, from, to)
}
\arguments{
\item{object}{A Ramanome object.}

\item{from}{The lower bound of the wavenumber range to cut.}

\item{to}{The upper bound of the wavenumber range to cut.}
}
\value{
A Ramanome object with truncated spectra, named as "cut.data" in the 'datasets' slot
}
\description{
Cutoff the spectra to the specific wavenumber range.
}
\examples{
data(RamEx_data)
Preprocessing.Cutoff(RamEx_data,550, 1800)
}
