% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{Preprocessing.Normalize}
\alias{Preprocessing.Normalize}
\title{Normalize the spectra matrix}
\usage{
Preprocessing.Normalize(object, method, wave = NULL)
}
\arguments{
\item{object}{A Ramanome object.}

\item{method}{The normalize method. "area" for area normalization, "specific" for specific wavenumber normalization, "max" for maximum normalization, "ch" for continuum normalization.}

\item{wave}{The interested wavenumber for specific normalization.}
}
\value{
A Ramanome object with normalized spectra, named as "normalized.data" in the 'datasets' slot
}
\description{
Normalize the spectra matrix
}
\examples{
data(RamEx_data)
data_normalized_ch <- Preprocessing.Normalize(RamEx_data, "ch")
#data_normalized_specific <- Preprocessing.Normalize(RamEx_data, "specific", 1650)
data_normalized_max <- Preprocessing.Normalize(RamEx_data, "max")
data_normalized_area <- Preprocessing.Normalize(RamEx_data, "area")
}
