% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{Preprocessing.OneStep}
\alias{Preprocessing.OneStep}
\title{One step spectral preprocessing}
\usage{
Preprocessing.OneStep(object)
}
\arguments{
\item{object}{Ramanome object}
}
\value{
The Ramanome object with preprocessed spectral data
}
\description{
This function performs one-step spectral preprocessing on the input Ramanome, including smoothing(SG), baseline correction(Polyfit), normalization(ch), and quality control(ICOD).
If specific preprocessing methods are needed, please define/run them step by step.
}
\examples{
data(RamEx_data)
data_preprocessed <- Preprocessing.OneStep(RamEx_data) 
}
