% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{Preprocessing.Smooth.Sg}
\alias{Preprocessing.Smooth.Sg}
\title{Smooth Spectral Data Using Savitzky-Golay Filter to reduce high-frequency noises}
\usage{
Preprocessing.Smooth.Sg(object, m = 0, p = 5, w = 11)
}
\arguments{
\item{object}{A Ramanome object.}

\item{m}{Whether perform differentiation. Defaults to 0.}

\item{p}{The order of the polynomial used in the Savitzky-Golay filter. Defaults to 5.}

\item{w}{The window size used for polynomial fitting. Defaults to 11.}
}
\value{
A Ramanome object with smoothed spectra, named as "smooth.data" in the 'datasets' slot
}
\description{
The Savitzky-Golay filter is a digital smoothing technique that preserves the shape and height of spectral peaks by fitting a polynomial to a moving window of data points.
}
\examples{
data(RamEx_data)
data_smoothed <- Preprocessing.Smooth.Sg(RamEx_data)
}
