% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocesssing.R
\name{Preprocesssing.Background.Remove}
\alias{Preprocesssing.Background.Remove}
\title{Perform background subtraction on Ramanome object}
\usage{
Preprocesssing.Background.Remove(object, cell.index, cal_mean = FALSE)
}
\arguments{
\item{object}{A Ramanome object.}

\item{cell.index}{The index of the cell component in the filenames.}

\item{cal_mean}{Logical value indicating whether to calculate the mean background spectrum.}
}
\value{
The modified Ramanome object.
}
\description{
This function performs background subtraction on a Ramanome object by subtracting the background spectra from the dataset.
}
\examples{
# Create a sample Ramanome object with background spectra
data <- matrix(rnorm(1000), nrow=10)
filenames <- c("group1_cell1", "group1_BG", "group2_cell1", "group2_BG")
wavenumbers <- seq(500, 3500, length.out=100)
meta <- data.frame(filenames=filenames, group=c("group1","group1","group2","group2"))
raman_obj <- new("Ramanome", datasets=list(raw.data=data), 
                 wavenumber=wavenumbers, meta.data=meta)


}
