% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocesssing.R
\name{Preprocesssing.Baseline.Bubble}
\alias{Preprocesssing.Baseline.Bubble}
\title{Remove baseline by bubble}
\usage{
Preprocesssing.Baseline.Bubble(object)
}
\arguments{
\item{object}{Ramanome object}
}
\value{
The baseline corrected spectra with bubble method
}
\description{
This function performs baseline pre-processing on the input object.
}
\examples{
data(RamEx_data)
data_smoothed <- Preprocesssing.Smooth.Sg(RamEx_data)
data_baseline <- Preprocesssing.Baseline.Polyfit(data_smoothed)
data_baseline_bubble <- Preprocesssing.Baseline.Bubble(data_smoothed)
}
