% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocesssing.R
\name{Preprocesssing.Normalize}
\alias{Preprocesssing.Normalize}
\title{Pre-process Raman data by ch normalization method}
\usage{
Preprocesssing.Normalize(object, method, wave = NULL)
}
\arguments{
\item{object}{A Ramanome object.}

\item{method}{The normalize method}

\item{wave}{The interested wavenumber}
}
\value{
The updated Ramanome object with normalized Raman data.
}
\description{
This function applies normalization to the Raman data based on the ch method/max method/specific method/area method and updates the Raman object.
}
\examples{
# Create a sample Ramanome object
data <- matrix(rnorm(1000), nrow=10)
wavenumbers <- seq(500, 3500, length.out=100)
raman_obj <- new("Ramanome", datasets=list(raw.data=data), wavenumber=wavenumbers)

# Apply different normalization methods
norm_area <- Preprocesssing.Normalize(raman_obj, method="area")
norm_max <- Preprocesssing.Normalize(raman_obj, method="max") 
norm_ch <- Preprocesssing.Normalize(raman_obj, method="ch")
norm_specific <- Preprocesssing.Normalize(raman_obj, method="specific", wave=1500)

}
