% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality_control_jump.R
\name{qc_jump}
\alias{qc_jump}
\title{Quality Control by Identifying Jumps in a Matrix}
\usage{
qc_jump(matrix, var_tol = 0.5, max_iterations = 100, kernel = c(1, 1, 1))
}
\arguments{
\item{matrix}{A numeric matrix containing the data to be analyzed.}

\item{var_tol}{A numeric value specifying the tolerance for variation in peak detection.
Defaults to 0.5.}

\item{max_iterations}{The maximum number of iterations to perform in the clustering process.
Defaults to 100.}

\item{kernel}{A numeric vector specifying the kernel to use in the convolution step.
Defaults to c(1,1,1).}
}
\value{
A list containing the time gaps, time clusters, index of good data points,
  and the iterations matrix.
}
\description{
This function performs quality control on a matrix by identifying jumps or outliers.
It uses a combination of bubble filling, peak detection, and clustering to identify
regions of the matrix that are significantly different from the rest.
}
