% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality-Control.R
\name{Qualitycontrol.All}
\alias{Qualitycontrol.All}
\title{Detect outliers in a dataset using various quality control methods}
\usage{
Qualitycontrol.All(matrix, var_tol = 0.5)
}
\arguments{
\item{matrix}{A matrix of spectral data where each row represents a single spectrum.}

\item{var_tol}{Tolerance for variation in the data. Defaults to 0.5.}
}
\value{
A list containing the outliers and the run time for each method.
}
\description{
This function applies multiple quality control methods to detect outliers in a dataset.
It uses Hotelling's T2, SNR, MCD, and distance methods to identify and mark outliers.
}
