% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality-Control.R
\name{Qualitycontrol.ICOD}
\alias{Qualitycontrol.ICOD}
\title{Quality Control by Identifying Jumps in a Matrix}
\usage{
Qualitycontrol.ICOD(
  object,
  var_tol = 1,
  max_iterations = 100,
  kernel = c(1, 1, 1)
)
}
\arguments{
\item{object}{Ramanome Object}

\item{var_tol}{A numeric value specifying the tolerance for variation of peaks in outliets detection.
Defaults to 0.5.}

\item{max_iterations}{The maximum number of iterations to perform in the clustering process.
Defaults to 100.}

\item{kernel}{A numeric vector specifying the kernel to use in the convolution step.
Defaults to c(1,1,1).}
}
\value{
A data frame with quality defined by ICOD method.
}
\description{
This function performs quality control on a matrix by identifying jumps or outliers.
It uses a combination of bubble filling, peak detection, and clustering to identify
regions of the matrix that are significantly different from the rest.
}
\examples{
data(RamEx_data)
qc_icod <- Qualitycontrol.ICOD(RamEx_data)
}
