% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality-Control.R
\name{Qualitycontrol.Mcd}
\alias{Qualitycontrol.Mcd}
\title{Detect outliers using the Minimum Covariance Determinant (MCD) method}
\usage{
Qualitycontrol.Mcd(x, h = 0.5, alpha = 0.01, na.rm = TRUE)
}
\arguments{
\item{x}{The input dataset.}

\item{h}{The fraction of the data to use for computing the MCD.}

\item{alpha}{The significance level for the MCD.}

\item{na.rm}{A logical value indicating whether to remove NA values.}
}
\value{
An object of class "Qualitycontrol.Mcd" containing the following components:
  - distance: The maximum distance threshold for outliers.
  - center: The center of the outliers.
  - outliers_pos: The positions of the outliers.
  - nb: A named number giving the total number of outliers detected.
}
\description{
This function extends the outliers_mcdEst function to detect outliers in a dataset
using the Minimum Covariance Determinant method. It modifies the output to include
additional information such as the distance and center of the outliers.
}
\examples{
data(RamEx_data)
data_smoothed <- Preprocessing.Smooth.Sg(RamEx_data)
data_baseline <- Preprocessing.Baseline.Polyfit(data_smoothed)
data_baseline_bubble <- Preprocessing.Baseline.Bubble(data_smoothed)
data_normalized <- Preprocessing.Normalize(data_baseline, "ch")
qc_icod <- Qualitycontrol.ICOD(data_normalized@datasets$normalized.data,var_tol = 0.4)
#qc_mcd <- Qualitycontrol.Mcd(data_normalized@datasets$normalized.data)
}
