% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality-Control.R
\name{Qualitycontrol.Snr}
\alias{Qualitycontrol.Snr}
\title{Calculate the Signal-to-Noise Ratio (SNR) for a dataset}
\usage{
Qualitycontrol.Snr(data, level = "medium")
}
\arguments{
\item{data}{A matrix of spectral data where each row represents a single spectrum.}

\item{level}{The level of strictness for the SNR calculation. Defaults to "medium".}
}
\value{
A vector indicating whether each spectrum passes the SNR test.
}
\description{
This function calculates the SNR for a dataset by comparing the signal and noise levels.
}
\examples{
data(RamEx_data)
data_smoothed <- Preprocesssing.Smooth.Sg(RamEx_data)
data_baseline <- Preprocesssing.Baseline.Polyfit(data_smoothed)
data_baseline_bubble <- Preprocesssing.Baseline.Bubble(data_smoothed)
data_normalized <- Preprocesssing.Normalize(data_baseline, "ch")
data_cleaned <- Qualitycontrol.ICOD(data_normalized@datasets$normalized.data,var_tol = 0.4)
qc_snr <- Qualitycontrol.Snr(data_normalized@datasets$normalized.data)
}
