% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality-Control.R
\name{Qualitycontrol.T2}
\alias{Qualitycontrol.T2}
\title{Detect outliers using Hotelling's T2 method}
\usage{
Qualitycontrol.T2(object, signif = 0.05)
}
\arguments{
\item{object}{Ramanome Object}

\item{signif}{Significance level (between 0 and 1), defaults to 0.05}
}
\value{
A data frame with quality defined by Hotelling's T2 method and their corresponding p-value
}
\description{
Detect outliers using Hotelling's T2 method
}
