% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantification.R
\name{Quantification.Glm}
\alias{Quantification.Glm}
\title{Generalized linear model (GLM)
SExtends linear regression by allowing the dependent
variable to follow distributions other than
normal (e.g., binomial, Poisson) and uses a link function
 to relate predictors to the response}
\usage{
Quantification.Glm(train, test = NULL)
}
\arguments{
\item{train}{The training data object}

\item{test}{The test data object (optional)}
}
\value{
The GLM prediction result.
}
\description{
Generalized linear model (GLM)
SExtends linear regression by allowing the dependent
variable to follow distributions other than
normal (e.g., binomial, Poisson) and uses a link function
 to relate predictors to the response
}
