% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{Raman.image}
\alias{Raman.image}
\title{Generate Raman imaging}
\usage{
Raman.image(object, peak)
}
\arguments{
\item{object}{The Ramanome object}

\item{peak}{The specified peak for Raman imaging}
}
\description{
This function generates a Raman imaging plot for a specified peak in a Ramanome object.
}
