% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raman-Markers.R
\name{Raman.Markers.DAPC}
\alias{Raman.Markers.DAPC}
\title{Find Markers Using Discriminant Analysis of Principal Components (DAPC)}
\usage{
Raman.Markers.DAPC(object, n_pc = 20, top_n = 100)
}
\arguments{
\item{object}{A Ramanome object.}

\item{n_pc}{Integer. The number of principal components to retain for LDA. Default is 20.}

\item{top_n}{Integer. For each group, how many top variables to return according to absolute importance. Default is 100.}
}
\value{
A data frame with the following columns:
\describe{
  \item{group}{The group label being discriminated}
  \item{wave}{The wavelength}
  \item{importance}{The calculated importance (projected LDA coefficient)}
}
Features of each group are sorted descendingly by absolute importance, and only the top N are reported.
}
\description{
DAPC is a statistical method for classifying groups and identifying features that discriminate between them.
Principal Component Analysis (PCA) is firstly used to reduce data dimensionality and remove correlations among variables.
Then for each group, a one-vs-rest Linear Discriminant Analysis (LDA) is performed on the selected PCs to find the linear combinations that best separate the predefined groups.
And the loadings are back-projected to the original feature space to trace variable importance. The top N features most important for discriminating each group are returned.
}
\details{
For each unique group label, a one-vs-rest LDA is trained on the PCA-reduced data, 
then the LDA coefficients are reconstructed back into the original feature space to rank feature importances.
}
\examples{
data(RamEx_data)
data_processed <- Preprocessing.OneStep(RamEx_data)
Markers_DAPC <- Raman.Markers.DAPC(preprocessing.cutoff(data_processed, 800, 1300), n_pc = 15, top_n = 40)
}
