% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raman-Markers.R
\name{Raman.Markers.Rbcs}
\alias{Raman.Markers.Rbcs}
\title{Perform random forest classification and feature importance analysis}
\usage{
Raman.Markers.Rbcs(
  object,
  outfolder = "RF_imps",
  threshold = 0.002,
  ntree = 1000,
  errortype = "oob",
  verbose = FALSE,
  nfolds = 3,
  rf.cv = FALSE,
  show = TRUE,
  save = FALSE
)
}
\arguments{
\item{object}{A Ramanome object.}

\item{outfolder}{The output folder to save the results.}

\item{threshold}{The threshold value for RBCS feature importances.}

\item{ntree}{The number of trees to build the random forest model.}

\item{errortype}{The error type used in random forest model evaluation.}

\item{verbose}{Logical value indicating whether to display progress messages.}

\item{nfolds}{The number of folds used in cross-validation.}

\item{rf.cv}{Logical value indicating whether to perform random forest with cross-validation.}

\item{show}{Logical value indicating whether to show the heatmap.}

\item{save}{Logical value indicating whether to draw the heatmap.}
}
\value{
A data frame containing the selected feature importances.
}
\description{
This function performs random forest classification on Raman data and generates feature importances.
}
\details{
more details about RBCS could be found in the reference below.
Teng L., Wang X.,  Wang X.,  Gou H.,  Ren L., & Wang T., 2016. Label-free, rapid and quantitative phenotyping of stress response in e. coli via ramanome. *Scientific Reports*
}
\examples{
data(RamEx_data)
data_processed <- Preprocessing.OneStep(RamEx_data)
Markers_RBCS <- Raman.Markers.Rbcs(data_processed, threshold = 0.003, show = TRUE)
}
