% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raman-Markers.R
\name{Raman.Markers.Roc}
\alias{Raman.Markers.Roc}
\title{Find Markers Using receiver operator characteristic (ROC) curve}
\usage{
Raman.Markers.Roc(
  object,
  threshold = 0.75,
  paired = FALSE,
  batch_size = 1000,
  n_threads = NULL
)
}
\arguments{
\item{object}{A Ramanome object.}

\item{threshold}{The minimum AUC threshold for considering a marker significant. Defaults to 0.75.}

\item{paired}{Whether to consider paired markers, e.g. wave1 / wave2. Default is FALSE}

\item{batch_size}{Sample size for each batch when traversing all possible paired markers. Default is 1,000}
}
\value{
A list containing:
\describe{
  \item{markers_single}{A data frame of single wavelength markers with their AUC scores}
  \item{markers_paired}{A data frame of paired wavelength markers with their AUC scores (only if paired=TRUE)}
}

A list containing two data frames: 'markers_single' for single markers and
'markers_paired' for paired markers, each with their corresponding AUC scores.
}
\description{
Performs receiver operator characteristic curve (ROC) analysis to identify significant markers by traversing all possible single and paired markers, and calculating the Area Under the Curve (AUC) for them, thus identify Raman features that best discriminate between different groups.
Each group/sample will get their own marker list.
}
\examples{
data(RamEx_data)
data_processed <- Preprocessing.OneStep(RamEx_data)
Markers_ROC <- Raman.Markers.Roc(preprocessing.cutoff(data_processed, 800, 1300), threshold = 0.75, paired = FALSE, batch_size = 1000)
}
