% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-new-class.R, R/base-ramanome-class.R
\docType{class}
\name{Ramanome-class}
\alias{Ramanome-class}
\title{Ramanome class}
\description{
This class represents the Ramanome data.

This class represents a Ramanome object, which contains Raman data and associated metadata.
}
\section{Slots}{

\describe{
\item{\code{datasets}}{A list of datasets}

\item{\code{meta.data}}{A data frame for meta data}

\item{\code{reductions}}{A list of reductions}

\item{\code{interested.bands}}{A list of interested bands}

\item{\code{wavenumber}}{Numeric vector indicating the wavenumbers of the Raman spectra}

\item{\code{datasets}}{A list of Raman datasets}

\item{\code{meta.data}}{A data frame containing metadata associated with the Raman datasets}

\item{\code{reductions}}{A list of data reduction results}

\item{\code{interested.bands}}{A list of interested Raman bands}
}}

