% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.r
\docType{data}
\name{RamEx.colors}
\alias{RamEx.colors}
\title{export RamEx.colors}
\format{
An object of class \code{character} of length 16.
}
\usage{
RamEx.colors
}
\description{
export RamEx.colors
}
\keyword{datasets}
