% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{RamEx_data}
\alias{RamEx_data}
\title{RamEx-data Dataset}
\format{
A Ramanome data
}
\usage{
data(RamEx_data)
}
\description{
The "RamEx-data" dataset originates from continuous sampling and monitoring during the liquid fermentation process of Saccharomyces cerevisiae. Samples were taken at the following time points:4, 12, 20, 28 and 36 hours. After three rounds of centrifugation, resuspension, and washing, the cells were air-dried naturally. The spectra were collected using a CAST-R device. The 532 nm laser power was set at 100 mW, with an integration time of 1 second. A 100× objective lens (NA = xx, Olympus, JP) was used to focus the laser beam on the sample.
}
\keyword{datasets}
