% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcs.R
\name{rbcs.rf}
\alias{rbcs.rf}
\title{Perform random forest classification and feature importance analysis}
\usage{
rbcs.rf(
  object,
  outfolder = "RF_imps",
  draw = TRUE,
  threshold = 0.002,
  ntree = 1000,
  errortype = "oob",
  verbose = FALSE,
  nfolds = 3,
  rf.cv = FALSE
)
}
\arguments{
\item{object}{A Ramanome object.}

\item{outfolder}{The output folder to save the results.}

\item{draw}{Logical value indicating whether to draw a heatmap.}

\item{threshold}{The threshold value for feature importances.}

\item{ntree}{The number of trees in the random forest model.}

\item{errortype}{The error type used in random forest model evaluation.}

\item{verbose}{Logical value indicating whether to display progress messages.}

\item{nfolds}{The number of folds used in cross-validation.}

\item{rf.cv}{Logical value indicating whether to perform random forest with cross-validation.}
}
\value{
A data frame containing the selected feature importances.
}
\description{
This function performs random forest classification on Raman data and generates feature importances.
}
