% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-ramanome-class.R
\name{rbind2,Ramanome,Ramanome-method}
\alias{rbind2,Ramanome,Ramanome-method}
\title{Combine Ramanome objects by row binding}
\usage{
\S4method{rbind2}{Ramanome,Ramanome}(x, y)
}
\arguments{
\item{x}{The first Ramanome object}

\item{y}{The second Ramanome object}
}
\value{
None
}
\description{
Combine Ramanome objects by row binding
}
