% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-spec.R
\name{read.single}
\alias{read.single}
\title{read single spectral data from cell path}
\usage{
read.single(cell_path, xout)
}
\arguments{
\item{cell_path}{The file path}

\item{xout}{A numeric vector with two values representing the lower and upper bound cutoff values.}
}
\value{
The subset of spectral data within the specified cutoff range.
}
\description{
This function cread single spectral data from cell path.
}
