% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-spec.R
\name{read.spec}
\alias{read.spec}
\title{Read spectral data from files}
\usage{
read.spec(
  data_path,
  group.index = 1,
  group.levels = NULL,
  cutoff = c(500, 3150),
  interpolation = FALSE
)
}
\arguments{
\item{data_path}{The path to the directory containing the spectral files.}

\item{group.index}{The index of the group information in the file names. Default is 1.}

\item{group.levels}{The levels of the group variable. Default is NULL.}

\item{cutoff}{A numeric vector with two values representing the lower and upper bound cutoff values for the wavenumber range. Default is c(500, 3150).}

\item{interpolation}{logical, should the wavelength axis be interpolated?}
}
\value{
A Ramanome object containing the spectral data, wavenumber information, and meta data.
}
\description{
This function reads spectral data from files and returns a Ramanome object.
}
