% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{reductions.pca}
\alias{reductions.pca}
\title{Perform PCA Reduction and Plotting}
\usage{
reductions.pca(object, draw = TRUE, save = FALSE)
}
\arguments{
\item{object}{A Ramanome object containing the dataset and metadata.}

\item{draw}{A logical value indicating whether to draw the PCA plot. Defaults to TRUE.}

\item{save}{A logical value indicating whether to save the plot as a file. Defaults to FALSE.}
}
\value{
The updated Ramanome object with the PCA results appended to the `reductions` slot.
}
\description{
This function performs Principal Component Analysis (PCA) on the dataset stored in
a Ramanome object and plots the first two principal components. The function can
also save the plot as a PNG file.
}
