% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{reductions}
\alias{reductions}
\title{Perform data reduction on Ramanome object}
\usage{
reductions(object, method, draw = FALSE)
}
\arguments{
\item{object}{A Ramanome object.}

\item{method}{The reduction method to use (e.g., "PCA", "tSNE", "UMAP").}

\item{draw}{Logical value to indicate whether to draw the reduction plot. Default is FALSE.}
}
\value{
The updated Ramanome object with the reduced data added to the @reductions slot.
}
\description{
This function performs data reduction on a Ramanome object using various methods such as PCA, t-SNE, or UMAP.
}
