% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{reductions.tsne}
\alias{reductions.tsne}
\title{Perform data reduction use tSNE method on Ramanome object}
\usage{
reductions.tsne(object, draw = FALSE)
}
\arguments{
\item{object}{A Ramanome object.}

\item{draw}{Logical value to indicate whether to draw the reduction plot. Default is FALSE.}
}
\value{
The updated Ramanome object with the reduced data added to the reductions slot.
}
\description{
This function performs data reduction on a Ramanome object using tSNE
}
