% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{reductions.umap}
\alias{reductions.umap}
\title{Perform data reduction use UMAP method on Ramanome object}
\usage{
reductions.umap(object, draw = FALSE)
}
\arguments{
\item{object}{A Ramanome object.}

\item{draw}{Logical value to indicate whether to draw the reduction plot. Default is FALSE.}
}
\value{
The updated Ramanome object with the reduced data added to the reductions slot.
}
\description{
This function performs data reduction on a Ramanome object using UMAP
}
