% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raman-Markers.R
\name{rf.out.of.bag}
\alias{rf.out.of.bag}
\title{Perform out-of-bag analysis for random forest classification}
\usage{
rf.out.of.bag(x, y, verbose = FALSE, ntree = 500, ...)
}
\arguments{
\item{x}{A matrix or data frame containing the predictors.}

\item{y}{A vector containing the response variable.}

\item{verbose}{A logical value indicating whether to print the progress of the model.}

\item{ntree}{An integer specifying the number of trees to grow.}

\item{...}{Additional arguments to be passed to the randomForest function.}
}
\value{
A list containing the following components:
\describe{
  \item{rf.model}{The random forest model object.}
  \item{probabilities}{The out-of-bag class probabilities for each observation.}
  \item{y}{The true response variable.}
  \item{predicted}{The predicted response variable based on the random forest model.}
  \item{confusion.matrix}{A confusion matrix showing the count of true and predicted levels.}
  \item{params}{A list of parameter values used for the random forest model.}
  \item{errs}{A vector of error rates for each observation.}
  \item{importances}{A vector of feature importances based on mean decrease accuracy.}
}
}
\description{
This function performs out-of-bag analysis for random forest classification on Raman data.
}
