% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-renishaw-data.R
\name{save.renishaw}
\alias{save.renishaw}
\title{Save Renishaw Raman data}
\usage{
save.renishaw(dir_path)
}
\arguments{
\item{dir_path}{The directory path where the Raman data is located.}
}
\description{
This function saves the Renishaw Raman data after processing.
}
