% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcs.R
\name{save.rf.results.mislabeling}
\alias{save.rf.results.mislabeling}
\title{Title}
\usage{
save.rf.results.mislabeling(result, filename = "mislabeling.xls", outdir)
}
\arguments{
\item{result}{The result object containing the predictions and probabilities. It should have at least 'y' (predicted labels) and 'probabilities' (predicted probabilities).}

\item{filename}{The name of the file to save the mislabeling results. Defaults to 'mislabeling.xls'.}

\item{outdir}{The directory where the output file will be saved.}
}
\value{
None. This function writes the mislabeling results to a file specified by 'filename' in the 'outdir' directory.
}
\description{
Title
}
\examples{
save.rf.results.mislabeling(result, 'mislabeling_results.xls', '/path/to/output/directory')
}
