% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcs.R
\name{save.rf.results.probabilities}
\alias{save.rf.results.probabilities}
\title{Save probabilities of random forest classification results}
\usage{
save.rf.results.probabilities(
  result,
  filename = "cv_probabilities.xls",
  outdir
)
}
\arguments{
\item{result}{The result object from random forest classification}

\item{filename}{The filename for the probabilities file (default is "cv_probabilities.xls")}

\item{outdir}{The output directory to save the probabilities file}
}
\description{
This function saves the probabilities of random forest classification results to a file.
}
