% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raman-Markers.R
\name{save.rf.results.summary}
\alias{save.rf.results.summary}
\title{Save summary of random forest classification results}
\usage{
save.rf.results.summary(result, rf.opts, filename = "summary.xls", outdir)
}
\arguments{
\item{result}{The result object from random forest classification}

\item{rf.opts}{The options used for random forest classification}

\item{filename}{The filename for the summary file (default is "summary.xls")}

\item{outdir}{The output directory to save the summary file}
}
\value{
None
}
\description{
This function saves a summary of the random forest classification results to a file.
}
\examples{
# Create sample results
set.seed(123)
result <- list(
  errs = rnorm(10, 0.2, 0.05),
  error.type = "oob",
  params = list(ntree = 500)
)
rf.opts <- list(errortype = "oob")

}
