% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamEx-Basic-Funciton.R
\name{select.band}
\alias{select.band}
\title{Select a range of Raman Feature.Reduction.Intensity values for a given wavenumber range}
\usage{
\method{select}{band}(object, waves)
}
\arguments{
\item{object}{A Ramanome object.}

\item{waves}{A vector with two values representing the lower and upper bounds of the wavenumber range.}
}
\value{
A numeric vector containing the summed Feature.Reduction.Intensity values across the specified wavenumber range for all spectra in the dataset.
}
\description{
This function selects a range of Raman Feature.Reduction.Intensity values for a given wavenumber range from a Ramanome object.
}
