% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamEx-Basic-Funciton.R
\name{select.value}
\alias{select.value}
\title{Select a single Raman Feature.Reduction.Intensity value for a given wavenumber}
\usage{
\method{select}{value}(object, wave)
}
\arguments{
\item{object}{A Ramanome object.}

\item{wave}{The wavenumber for which to retrieve the Feature.Reduction.Intensity value.}
}
\value{
A numeric vector containing the Feature.Reduction.Intensity values at the specified wavenumber for all spectra in the dataset.
}
\description{
This function selects a single Raman Feature.Reduction.Intensity value for a given wavenumber from a Ramanome object.
}
\examples{
# Create a sample Ramanome object
wavenumbers <- seq(500, 3500, by = 10)
spectra <- matrix(rnorm(100 * length(wavenumbers)), nrow = 100)
raman_obj <- new("Ramanome",
  datasets = list(raw = spectra),
  wavenumber = wavenumbers
)

}
