% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality_control_Other_methods.R
\name{spc_QC_SNR}
\alias{spc_QC_SNR}
\title{Calculate the Signal-to-Noise Ratio (SNR) for a dataset}
\usage{
spc_QC_SNR(data, level = "medium")
}
\arguments{
\item{data}{A matrix of spectral data where each row represents a single spectrum.}

\item{level}{The level of strictness for the SNR calculation. Defaults to "medium".}
}
\value{
A vector indicating whether each spectrum passes the SNR test.
}
\description{
This function calculates the SNR for a dataset by comparing the signal and noise levels.
}
