% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-mean-spec.R
\name{spec.mean.draw}
\alias{spec.mean.draw}
\title{Draw mean spectrum}
\usage{
spec.mean.draw(object, gap = 0)
}
\arguments{
\item{object}{The Ramanome object}

\item{gap}{The gap between the mean spectra (default: 0)}
}
\value{
The plot of the mean spectrum
}
\description{
This function generates a plot of the mean spectrum for a Ramanome object.
}
