% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-colored-by-peak.R
\name{special_peak.coloring}
\alias{special_peak.coloring}
\title{Color cells by special peak}
\usage{
special_peak.coloring(object, peak, reduction = "UMAP")
}
\arguments{
\item{object}{The Ramanome object}

\item{peak}{The special peak to color cells by}

\item{reduction}{The reduction method to use (default: 'UMAP')}
}
\value{
The plot object
}
\description{
This function colors cells in a Ramanome object by a special peak in a specified reduction.
}
