% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{stratified_partition}
\alias{stratified_partition}
\title{Perform stratified partitioning of data indices}
\usage{
stratified_partition(labels, p = 0.7)
}
\arguments{
\item{labels}{A vector of class labels}

\item{p}{The proportion of data to include in the training set (default: 0.7)}
}
\value{
A vector of indices for the training set
}
\description{
This function performs stratified sampling to split data indices into training and test sets
while maintaining class proportions.
}
