% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-series.R
\name{time.series}
\alias{time.series}
\title{Generate time series plots for Raman data}
\usage{
\method{time}{series}(object, reduction = "UMAP")
}
\arguments{
\item{object}{A Ramanome object.}

\item{reduction}{The reduction method to use (default is "UMAP").}
}
\value{
A combined ggplot object of the time series plots.
}
\description{
This function generates time series plots for Raman data using a specified reduction method (e.g., UMAP).
}
