% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irca.R
\name{vectorize_dm}
\alias{vectorize_dm}
\title{Vectorize Distance Matrix}
\usage{
vectorize_dm(dm, group = NULL, duplicate = TRUE)
}
\arguments{
\item{dm}{A distance matrix.}

\item{group}{A vector specifying the sample categories. (default is NULL)}

\item{duplicate}{A logical value indicating whether to include duplicate pairs. (default is TRUE)}
}
\value{
A data frame containing the vectorized distance matrix.
}
\description{
This function vectorizes a distance matrix, optionally grouping the samples by a metadata category.
}
