% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality_control_jump.R
\name{bubbleloop}
\alias{bubbleloop}
\title{Process a spectrum with bubble loop}
\usage{
bubbleloop(spectrum, baseline, min_bubble_widths)
}
\arguments{
\item{spectrum}{A numeric vector representing the spectrum.}

\item{baseline}{A numeric vector representing the baseline of the spectrum.}

\item{min_bubble_widths}{A numeric value or vector specifying the minimum width of the bubbles.}
}
\value{
A list containing the adjusted baseline and the bounds of each bubble.
}
\description{
This function iteratively applies the bubble growing algorithm to a spectrum, adjusting the baseline
and identifying the bounds of each bubble.
}
