% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{Classification.Gmm}
\alias{Classification.Gmm}
\title{Gaussian Mixture Model (GMM)}
\usage{
Classification.Gmm(train, test = NULL)
}
\arguments{
\item{train}{The training data object}

\item{test}{The test data object (optional)}
}
\value{
The GMM model.
}
\description{
A probabilistic model that assumes data is generated from
a mixture of Gaussian distributions and assigns probabilities
 to each class
}
\examples{
data(RamEx_data)
data_smoothed <- Preprocessing.Smooth.Sg(RamEx_data)
data_baseline <- Preprocessing.Baseline.Polyfit(data_smoothed)
data_normalized <- Preprocessing.Normalize(data_baseline, "ch")
qc_icod <- Qualitycontrol.ICOD(data_normalized@datasets$normalized.data,var_tol = 0.5)
data_cleaned <- RamEx_data[qc_icod$quality,]
data_cleaned <- Feature.Reduction.Intensity(data_cleaned, list(c(2000,2250),c(2750,3050), 1450, 1665))
Classification.Gmm(data_cleaned)
}
