% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{Classification.Gmm}
\alias{Classification.Gmm}
\title{Gaussian Mixture Model (GMM)}
\usage{
Classification.Gmm(train, test = NULL)
}
\arguments{
\item{train}{The training data object}

\item{test}{The test data object (optional). If not provided, the function will perform a stratified cross-validation (70%% for training and 30%% for testing of the input 'train').}
}
\value{
A list containing:
\describe{ 
  \item{model}{The GMM model}
  \item{pred_test}{The prediction for test data if test is provided}
}
}
\description{
A probabilistic model that assumes data is generated from
a mixture of Gaussian distributions and assigns probabilities
 to each class
}
\examples{
data(RamEx_data)
Classification.Gmm(RamEx_data)

}
