% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{Classification.Gmm}
\alias{Classification.Gmm}
\title{Gaussian Mixture Model (GMM)}
\usage{
Classification.Gmm(
  train,
  test = NULL,
  n_pc = 20,
  show = TRUE,
  save = FALSE,
  seed = 42
)
}
\arguments{
\item{train}{The training data object}

\item{test}{The test data object (optional). If not provided, the function will perform a stratified cross-validation (Training : Test = 7 : 3).}

\item{n_pc}{The number of principal components to use}
}
\value{
A list containing:
\describe{ 
  \item{model}{The GMM model}
  \item{pred_test}{The prediction for test data if test is provided}
}
}
\description{
A probabilistic model that assumes data is generated from
a mixture of Gaussian distributions and assigns probabilities
 to each class
}
\examples{
data(RamEx_data)
data_processed <- Preprocessing.OneStep(RamEx_data)
model.gmm <- Classification.Gmm(data_processed)
}
