% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{Classification.Lda}
\alias{Classification.Lda}
\title{Perform classification use linear discriminant analysis LDA model}
\usage{
Classification.Lda(
  train,
  test = NULL,
  show = TRUE,
  save = FALSE,
  seed = 42,
  n_pc = 20
)
}
\arguments{
\item{train}{The training data object}

\item{test}{The test data object (optional). If not provided, the function will perform a stratified cross-validation (Training : Test = 7 : 3).}

\item{show}{Whether user want to show the confusion matrix plot of the results}

\item{save}{Wether user want to save the confusion matrix plot of the results (default path : getwd())}

\item{seed}{The random seed}

\item{n_pc}{The number of principal components to use}
}
\value{
A list containing:
\describe{
  \item{model}{The LDA model}
  \item{pred_test}{The prediction for test data if test is provided}
}
}
\description{
This function performs classification using PC-LDA
}
\examples{
data(RamEx_data)
Classification.Lda(RamEx_data)
}
