% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{Classification.Lda}
\alias{Classification.Lda}
\title{Perform classification use LDA model}
\usage{
Classification.Lda(train, test = NULL, show = TRUE, save = FALSE, seed = 42)
}
\arguments{
\item{train}{The training data object}

\item{test}{The test data object (optional)}

\item{show}{Whether user want to show the results}

\item{save}{Wether user want to save the results}

\item{seed}{The random seed}
}
\value{
A list containing:
  \item{pred.train}{The confusion matrix plot for training data}
  \item{pred.test}{The confusion matrix plot for test data}
}
\description{
This function performs classification using PC-LDA
}
\examples{
data(RamEx_data)
data_smoothed <- Preprocesssing.Smooth.Sg(RamEx_data)
data_baseline <- Preprocesssing.Baseline.Polyfit(data_smoothed)
data_baseline_bubble <- Preprocesssing.Baseline.Bubble(data_smoothed)
data_normalized <- Preprocesssing.Normalize(data_baseline, "ch")
data_cleaned <- Qualitycontrol.ICOD(data_normalized@datasets$normalized.data,var_tol = 0.4)
data_cleaned <- data_normalized[data_cleaned$index_good,]
data_cleaned <- Feature.Reduction.Intensity(data_cleaned, list(c(2000,2250),c(2750,3050), 1450, 1665))
Classification.Lda(data_cleaned)
}
