% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification}
\alias{classification}
\title{Perform classification}
\usage{
classification(
  train,
  test = NULL,
  method = "PC-LDA",
  show = TRUE,
  save = FALSE,
  seed = 42
)
}
\arguments{
\item{train}{The training data object}

\item{test}{The test data object (optional)}

\item{method}{The classification method to use (default: 'PC-LDA')}

\item{show}{Whether user want to show the results}

\item{save}{Wether user want to save the results}

\item{seed}{The random seed}
}
\description{
This function performs classification using various methods such as PC-LDA, SVM, and Random Forest.
}
