% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{Classification.Rf}
\alias{Classification.Rf}
\title{Perform classification use RandomForest model}
\usage{
Classification.Rf(train, test = NULL, show = TRUE, save = FALSE, seed = 42)
}
\arguments{
\item{train}{The training data object}

\item{test}{The test data object (optional)}

\item{show}{Whether user want to show the results}

\item{save}{Wether user want to save the results}

\item{seed}{The random seed}
}
\value{
A list containing:
  \item{pred.train}{The confusion matrix plot for training data}
  \item{pred.test}{The confusion matrix plot for test data}
}
\description{
This function performs classification using RF
}
\examples{
data(RamEx_data)
data_smoothed <- Preprocessing.Smooth.Sg(RamEx_data)
data_baseline <- Preprocessing.Baseline.Polyfit(data_smoothed)
data_baseline_bubble <- Preprocessing.Baseline.Bubble(data_smoothed)
data_normalized <- Preprocessing.Normalize(data_baseline, "ch")
data_cleaned <- Qualitycontrol.ICOD(data_normalized@datasets$normalized.data,var_tol = 0.4)
data_cleaned <- RamEx_data[qc_icod$quality,]
data_cleaned <- Feature.Reduction.Intensity(data_cleaned, list(c(2000,2250),c(2750,3050), 1450, 1665))
Classification.Rf(data_cleaned)
}
