% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{Classification.Rf}
\alias{Classification.Rf}
\title{Perform classification use RandomForest (RF) model}
\usage{
Classification.Rf(
  train,
  test = NULL,
  ntree = 100,
  mtry = 2,
  show = TRUE,
  save = FALSE,
  seed = 42
)
}
\arguments{
\item{train}{The training data object}

\item{test}{The test data object (optional). If not provided, the function will perform a stratified cross-validation (Training : Test = 7 : 3).}

\item{ntree}{The number of trees in the forest}

\item{mtry}{The number of variables randomly sampled as candidates for splitting at each node}

\item{show}{Whether user want to show the confusion matrix plot of the results}

\item{save}{Wether user want to save the confusion matrix plot of the results (default path : getwd())}

\item{seed}{The random seed}
}
\value{
A list containing:
\describe{ 
  \item{model}{The RF model}
  \item{pred_test}{The prediction for test data if test is provided}
}
}
\description{
This function performs classification using RF
}
\examples{
data(RamEx_data)
data_processed <- Preprocessing.OneStep(RamEx_data)
model.rf <- Classification.Rf(data_processed)
}
