% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{Classification.Svm}
\alias{Classification.Svm}
\title{Perform classification use support vector machine (SVM) model}
\usage{
Classification.Svm(train, test = NULL, show = TRUE, save = FALSE, seed = 42)
}
\arguments{
\item{train}{The training data object}

\item{test}{The test data object (optional). If not provided, the function will perform a stratified cross-validation (Training : Test = 7 : 3).}

\item{show}{Whether user want to show the confusion matrix plot of the results}

\item{save}{Wether user want to save the confusion matrix plot of the results (default path : getwd())}

\item{seed}{The random seed}
}
\value{
A list containing:
\describe{
  \item{model}{The SVM model}
  \item{pred_test}{The prediction for test data if test is provided}
}
}
\description{
This function performs classification using SVM
}
\examples{
data(RamEx_data)
Classification.Svm(RamEx_data)
}
