% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamEx-Basic-Funciton.R
\name{cluster.color}
\alias{cluster.color}
\title{Generate a color-coded plot for cluster visualization}
\usage{
cluster.color(data, group)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{group}{The group to focus on.}
}
\value{
A ggplot object.
}
\description{
This function generates a color-coded plot to visualize clusters in a dataset.
}
