% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamEx-Basic-Funciton.R
\name{confirm.name}
\alias{confirm.name}
\title{string. If two values are provided, they are combined into a range string with a
tilde (~) separator. If more than two values are provided, the function stops with
an error.}
\usage{
confirm.name(waves)
}
\arguments{
\item{waves}{A single numeric value or a vector of two numeric values representing
wavelengths or a wavelength range.}
}
\value{
A character string representing the formatted wavelength name or range.
}
\description{
string. If two values are provided, they are combined into a range string with a
tilde (~) separator. If more than two values are provided, the function stops with
an error.
}
