% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_value.R
\name{confirm.select}
\alias{confirm.select}
\title{Confirm Selection and Retrieve Spectral Data}
\usage{
confirm.select(object, waves)
}
\arguments{
\item{object}{A Ramanome object containing spectral data and wavelength information.}

\item{waves}{A single numeric value representing a wavelength or a vector of two
numeric values representing a wavelength range.}
}
\value{
The selected spectral data based on the input wavelength or range.
}
\description{
This function determines the type of selection based on the input and retrieves the
corresponding spectral data from a Ramanome object. If a single wavelength is
provided, it uses `select.value` to retrieve the data at that wavelength. If a
range of wavelengths is provided, it uses `select.band` to retrieve the sum of
the data within that range.
}
\seealso{
select.value for retrieving data at a single wavelength.

select.band for retrieving the sum of data within a wavelength range.
}
