% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamEx-Basic-Funciton.R
\name{confusion.cal}
\alias{confusion.cal}
\title{Calculate accuracy of prediction matrix}
\usage{
confusion.cal(pred_matrix)
}
\arguments{
\item{pred_matrix}{The prediction matrix}
}
\value{
The accuracy matrix
}
\description{
This function calculates the accuracy of a prediction matrix.
}
