% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamEx-Basic-Funciton.R
\name{confusion.plot}
\alias{confusion.plot}
\title{Plot prediction results}
\usage{
confusion.plot(true_labels, pred_labels)
}
\arguments{
\item{true_labels}{The true labels}

\item{pred_labels}{The predicted labels}
}
\value{
A tile plot of the prediction results
}
\description{
This function plots the prediction results using a tile plot.
}
