% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamEx-Basic-Funciton.R
\name{cut.spec}
\alias{cut.spec}
\title{Cut Spectral Data Based on Wavelength Range}
\usage{
\method{cut}{spec}(single_spec, cutoff)
}
\arguments{
\item{single_spec}{A data frame containing spectral data with wavelengths in the first column (V1).}

\item{cutoff}{A numeric vector of two values specifying the lower and upper bounds of the wavelength range.}
}
\value{
A data frame containing the subset of the input spectral data where the wavelength values (V1)
        fall within the specified cutoff range.
}
\description{
This function extracts a subset of spectral data based on a specified wavelength range.
It filters the data to include only the rows where the wavelength (V1) is within the
lower and upper bounds of the cutoff range.
}
