% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature-Reduction.R
\name{Feature.Reduction.Intensity}
\alias{Feature.Reduction.Intensity}
\title{Extract Intensity Values at Specified Wavelengths}
\usage{
Feature.Reduction.Intensity(object, wavenumber)
}
\arguments{
\item{object}{A Ramanome object containing spectral data and wavelength information.}

\item{wavenumber}{A single numeric value, a vector of numeric values, or a list of
numeric values representing wavelengths or wavelength ranges.}
}
\value{
The updated Ramanome object with the extracted Feature.Reduction.Intensity values appended to
the `interested.bands` slot.
}
\description{
This function retrieves the Feature.Reduction.Intensity values at specified wavelengths or wavelength
ranges from a Ramanome object. It uses `confirm.select` to get the spectral data and
`confirm.name` to format the names of the selected wavelengths or ranges.
}
\examples{
# Load example data
data_file <- system.file("extdata", "data", package = "RamEx")
RamEx_data <- read.spec.load(data_file, group.index = 2)

# Extract intensity values at specific wavelengths
RamEx_data <- Feature.Reduction.Intensity(RamEx_data, c(1000, 1500))
}
