% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature-Reduction.R
\name{Feature.Reduction.Pca}
\alias{Feature.Reduction.Pca}
\title{Perform PCA Reduction and Plotting}
\usage{
Feature.Reduction.Pca(object, draw = TRUE, save = FALSE, n_pc = 2)
}
\arguments{
\item{object}{A Ramanome object containing the dataset and metadata.}

\item{draw}{A logical value indicating whether to draw the PCA plot. Defaults to TRUE.}

\item{save}{A logical value indicating whether to save the plot as a file. Defaults to FALSE.}

\item{n_pc}{The number of principal components to save. Defaults to 2.}
}
\value{
The updated Ramanome object with the PCA results appended to the `reductions` slot.
}
\description{
Performs Principal Component Analysis (PCA) based on the last spectral matrix in the 'datasets' slot
}
\examples{
data(RamEx_data)
data_processed <- Preprocessing.OneStep(RamEx_data)
data.reduction.pca <- Feature.Reduction.Pca(data_processed, draw=TRUE, save = FALSE)
}
