% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature-Reduction.R
\name{Feature.Reduction.Pcoa}
\alias{Feature.Reduction.Pcoa}
\title{Perform Principal Coordinate Analysis (PCoA) and Plot Results}
\usage{
Feature.Reduction.Pcoa(
  object,
  PCA = 20,
  n_pc = 2,
  distance = "euclidean",
  show = TRUE,
  save = FALSE
)
}
\arguments{
\item{object}{A Ramanome object containing the dataset and metadata.}

\item{PCA}{A numeric value indicating the number of principal components to use. Defaults to 20 (if NULL,raw data matrix will be used to generate PCoA plot).}

\item{n_pc}{The number of PCoA components to save. Defaults to 2.}

\item{distance}{The distance method to use. Defaults to "euclidean". Other methods see ?stats::dist.}

\item{show}{A logical value indicating whether to draw the PCoA plot. Defaults to TRUE.}

\item{save}{A logical value indicating whether to save the plot as a PNG file. Defaults to FALSE.}
}
\value{
The updated Ramanome object with the PCoA results appended to the `reductions` slot.
}
\description{
This function performs Principal Coordinate Analysis (PCoA) on the dataset stored in
the Ramanome object. It calculates the distance matrix using the 'euclidean' method
and then applies PCoA to reduce the dimensionality to two principal components.
If drawing is enabled, it creates a plot of the PCoA results, colored by group.
The plot can also be saved as a PNG file if specified.
}
\examples{
data(RamEx_data)
data_processed <- Preprocessing.OneStep(RamEx_data)
data.reduction.pcoa <- Feature.Reduction.Pcoa(data_processed, show=TRUE, save = FALSE)
}
