% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature-Reduction.R
\name{Feature.Reduction.Pcoa}
\alias{Feature.Reduction.Pcoa}
\title{Perform Principal Coordinate Analysis (PCoA) and Plot Results}
\usage{
Feature.Reduction.Pcoa(object, draw = TRUE, save = FALSE)
}
\arguments{
\item{object}{A Ramanome object containing the dataset and metadata.}

\item{draw}{A logical value indicating whether to draw the PCoA plot. Defaults to TRUE.}

\item{save}{A logical value indicating whether to save the plot as a PNG file. Defaults to FALSE.}
}
\value{
The updated Ramanome object with the PCoA results appended to the `reductions` slot.
}
\description{
This function performs Principal Coordinate Analysis (PCoA) on the dataset stored in
the Ramanome object. It calculates the distance matrix using the 'euclidean' method
and then applies PCoA to reduce the dimensionality to two principal components.
If drawing is enabled, it creates a plot of the PCoA results, colored by group.
The plot can also be saved as a PNG file if specified.
}
\examples{
# Load example data
data_file <- system.file("extdata", "data", package = "RamEx")
RamEx_data <- read.spec.load(data_file, group.index = 2)
}
