% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature-Reduction.R
\name{Feature.Reduction.Umap}
\alias{Feature.Reduction.Umap}
\title{Perform UMAP Dimensionality Reduction}
\usage{
Feature.Reduction.Umap(object, draw = TRUE, save = FALSE)
}
\arguments{
\item{object}{A Ramanome object containing the dataset.}

\item{draw}{A logical value indicating whether to generate a plot. Defaults to TRUE.}

\item{save}{A logical value indicating whether to save the plot as a PNG file. Defaults to FALSE.}
}
\value{
The updated Ramanome object with the UMAP reduction results.
}
\description{
This function performs Uniform Manifold Approximation and Projection (UMAP) on the
dataset stored in a Ramanome object. It first applies PCA to reduce the dataset
to 20 dimensions and then applies UMAP to reduce it to two dimensions. The reduced
data is stored in the object and a plot is generated if specified.
}
