% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcs.R
\name{get.mislabel.scores}
\alias{get.mislabel.scores}
\title{Calculate mislabel scores for classification results}
\usage{
get.mislabel.scores(y, y.prob)
}
\arguments{
\item{y}{A vector containing the true labels.}

\item{y.prob}{A matrix containing class probabilities.}
}
\value{
A matrix containing mislabel scores for each class.
}
\description{
This function calculates mislabel scores based on class probabilities for classification results.
}
