% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamEx-Basic-Funciton.R
\name{get.nearest.dataset}
\alias{get.nearest.dataset}
\title{Retrieve the final spectral matrix from a Ramanome Object}
\usage{
get.nearest.dataset(object)
}
\arguments{
\item{object}{A Ramanome object.}
}
\value{
A matrix or data frame containing the most recently added dataset from the Ramanome object.
}
\description{
This function extracts the most recently added dataset from a Ramanome object.
It is assumed that the dataset is stored in the `datasets` slot of the Ramanome object.
}
\examples{
data(RamEx_data)
data_processed <- Preprocessing.OneStep(RamEx_data)
data_matrix <- get.nearest.dataset(data_processed)
}
