% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamEx-Basic-Funciton.R
\name{get.nearest.dataset}
\alias{get.nearest.dataset}
\title{Retrieve the Nearest Dataset from a Ramanome Object}
\usage{
get.nearest.dataset(object)
}
\arguments{
\item{object}{A Ramanome object that contains datasets in its `datasets` slot.}
}
\value{
The most recently added dataset from the Ramanome object.
}
\description{
This function extracts the most recently added dataset from a Ramanome object.
It is assumed that the dataset is stored in the `datasets` slot of the Ramanome object.
}
