% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_value.R
\name{intensity}
\alias{intensity}
\title{Extract Intensity Values at Specified Wavelengths}
\usage{
intensity(object, wavenumber)
}
\arguments{
\item{object}{A Ramanome object containing spectral data and wavelength information.}

\item{wavenumber}{A single numeric value, a vector of numeric values, or a list of
numeric values representing wavelengths or wavelength ranges.}
}
\value{
The updated Ramanome object with the extracted intensity values appended to
the `interested.bands` slot.
}
\description{
This function retrieves the intensity values at specified wavelengths or wavelength
ranges from a Ramanome object. It uses `confirm.select` to get the spectral data and
`confirm.name` to format the names of the selected wavelengths or ranges.
}
