% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intra-Ramanome-Analysis.R
\name{Intraramanome.Analysis.2Dcos}
\alias{Intraramanome.Analysis.2Dcos}
\title{2D Correlation Spectroscopy Analysis (2D-COS)
Captures both synchronous (simultaneous changes) and
asynchronous (sequential changes) relationships, providing
detailed insights into spectral changes}
\usage{
Intraramanome.Analysis.2Dcos(object)
}
\arguments{
\item{object}{A Ramanome object containing the dataset.}
}
\value{
synchronous and asynchronous correlation spectra result
}
\description{
2D Correlation Spectroscopy Analysis (2D-COS)
Captures both synchronous (simultaneous changes) and
asynchronous (sequential changes) relationships, providing
detailed insights into spectral changes
}
\examples{
data(RamEx_data)
data_smoothed <- Preprocesssing.Smooth.Sg(RamEx_data)
data_baseline <- Preprocesssing.Baseline.Polyfit(data_smoothed)
data_baseline_bubble <- Preprocesssing.Baseline.Bubble(data_smoothed)
data_normalized <- Preprocesssing.Normalize(data_baseline, "ch")
data_cleaned <- Qualitycontrol.ICOD(data_normalized@datasets$normalized.data,var_tol = 0.4)
data_cleaned <- data_normalized[data_cleaned$index_good,]
data_cleaned <- Feature.Reduction.Intensity(data_cleaned, list(c(2000,2250),c(2750,3050), 1450, 1665))
#reslut_2dos <- Intraramanome.Analysis.2Dcos(data_cleaned)
}
