% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intra-Ramanome-Analysis.R
\name{Intraramanome.Analysis.Irca.Global.draw}
\alias{Intraramanome.Analysis.Irca.Global.draw}
\title{Perform IRCA analysis on global scale}
\usage{
Intraramanome.Analysis.Irca.Global.draw(dataset, group, threshold = 0.6)
}
\arguments{
\item{dataset}{The input dataset for analysis}

\item{threshold}{The threshold value for correlation. Default is 0.6.}

\item{save}{Logical value to indicate whether to save the Global IRCA plot. Default is FALSE.}
}
\value{
A dataframe containing the significant correlations between variables
}
\description{
This function performs IRCA analysis on the global scale using the input dataset and group information.
It generates a chord diagram that represents the significant correlations between variables.
}
