% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intra-Ramanome-Analysis.R
\name{Intraramanome.Analysis.Irca.Global}
\alias{Intraramanome.Analysis.Irca.Global}
\title{Perform global IRCA analysis}
\usage{
Intraramanome.Analysis.Irca.Global(object)
}
\arguments{
\item{object}{The Raman spectroscopy data object.}
}
\value{
A matrix containing the global IRCA interests.
}
\description{
This function performs global IRCA analysis on the given Raman spectroscopy data object.
It generates negative IRCA plots for each group and saves them as jpeg files.
}
\examples{
data(RamEx_data)
data_smoothed <- Preprocessing.Smooth.Sg(RamEx_data)
data_baseline <- Preprocessing.Baseline.Polyfit(data_smoothed)
data_normalized <- Preprocessing.Normalize(data_baseline, "ch")
qc_icod <- Qualitycontrol.ICOD(data_normalized@datasets$normalized.data,var_tol = 0.4)
data_cleaned <- data_normalized[qc_icod$quality,]
IRCA.interests <- Intraramanome.Analysis.Irca.Global(data_cleaned)
}
