% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rammanome-class.R
\name{length,Ramanome-method}
\alias{length,Ramanome-method}
\title{Length method for Ramanome objects}
\usage{
\S4method{length}{Ramanome}(x)
}
\arguments{
\item{x}{The Ramanome object}
}
\description{
Length method for Ramanome objects
}
