% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamEx-Basic-Funciton.R
\name{mean.spec}
\alias{mean.spec}
\title{Calculate Mean and Standard Deviation of Spectral Data by Group}
\usage{
\method{mean}{spec}(data, group, gap = 0.3)
}
\arguments{
\item{data}{A matrix or data frame containing the spectral data.}

\item{group}{A factor or character vector indicating the group for each row in the data.}

\item{gap}{A numeric value representing the vertical gap between the mean lines of different groups.}
}
\value{
A ggplot2 object containing:
  \item{Lines}{Mean spectra for each group}
  \item{Ribbons}{Standard deviation bands around each mean spectrum}
  \item{x-axis}{Wavenumber values}
  \item{y-axis}{Normalized intensity values}
  \item{Legend}{Group identifiers}
}
\description{
This function calculates the mean and standard deviation of spectral data for each group.
It then creates a ggplot2 plot with ribbons representing the standard deviation and lines
representing the mean for each group.
}
\examples{
data(RamEx_data)
data_smoothed <- Preprocesssing.Smooth.Sg(RamEx_data)
data_baseline <- Preprocesssing.Baseline.Polyfit(data_smoothed)
data_baseline_bubble <- Preprocesssing.Baseline.Bubble(data_smoothed)
data_normalized <- Preprocesssing.Normalize(data_baseline, "ch")
mean.spec(data_normalized@datasets$baseline.data, data_normalized@meta.data$group)
}
