% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamEx-Basic-Funciton.R
\name{mean.spec}
\alias{mean.spec}
\title{Calculate Mean and Standard Deviation of Spectral Data by Group and Plot}
\usage{
\method{mean}{spec}(data, group, gap = 0.3, col = NULL)
}
\arguments{
\item{data}{A matrix or data frame containing the spectral data, colnames are wavenumbers.}

\item{group}{Factor indicating the group for each row in the data.}

\item{gap}{A numeric value representing the vertical gap between the mean lines of different groups.}

\item{col}{A vector of colors for each group. If NULL, the default colors will be used.}
}
\value{
A ggplot2 plot
}
\description{
Calculates the mean and standard deviation of spectral data for each group then creates a plot with ribbons representing the standard deviation and lines
}
\examples{
data(RamEx_data)
data_processed <- Preprocessing.OneStep(RamEx_data)
mean.spec(data_processed@datasets$normalized.data, data_processed@meta.data$group)
}
