% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-mean-spec.R
\name{mean.spec}
\alias{mean.spec}
\title{Calculate Mean and Standard Deviation of Spectral Data by Group}
\usage{
\method{mean}{spec}(data, group, gap = 0.3)
}
\arguments{
\item{data}{A matrix or data frame containing the spectral data.}

\item{group}{A factor or character vector indicating the group for each row in the data.}

\item{gap}{A numeric value representing the vertical gap between the mean lines of different groups.}
}
\value{
A ggplot object representing the plot.
}
\description{
This function calculates the mean and standard deviation of spectral data for each group.
It then creates a ggplot2 plot with ribbons representing the standard deviation and lines
representing the mean for each group.
}
