% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotype-Analysis.R
\name{Phenotype.Analysis.Hca}
\alias{Phenotype.Analysis.Hca}
\title{Hierarchical Clustering Analysis (HCA)}
\usage{
Phenotype.Analysis.Hca(object)
}
\arguments{
\item{object}{A Ramanome object.}
}
\value{
The hierarchical clustering fit object.
}
\description{
An iterative clustering method that builds a nested tree of clusters by progressively merging similar samples or dividing dissimilar ones, visualized as a dendrogram showing relationships at multiple scales.
The function also plots the resulting dendrogram.
}
\examples{
data(RamEx_data)
data_processed <- Preprocessing.OneStep(RamEx_data)
clusters_hca <- Phenotype.Analysis.Hca(data_processed)
}
