% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotype-Analysis.R
\name{Phenotype.Analysis.Hca}
\alias{Phenotype.Analysis.Hca}
\title{Perform Hierarchical Clustering Analysis (HCA)}
\usage{
Phenotype.Analysis.Hca(object)
}
\arguments{
\item{object}{A Ramanome object containing the dataset.}
}
\value{
The hierarchical clustering fit object.
}
\description{
This function performs hierarchical clustering on the dataset retrieved from a
Ramanome object using the average linkage method. It calculates the Euclidean distance
matrix and then applies the hierarchical clustering algorithm to create a dendrogram.
The function also plots the resulting dendrogram.
}
\examples{
# Load example Ramanome data
data_file <- system.file("extdata", "data", package = "RamEx")
RamEx_data <- read.spec.load(data_file, group.index = 2)

}
