% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotype-Analysis.R
\name{Phenotype.Analysis.Kmeans}
\alias{Phenotype.Analysis.Kmeans}
\title{k-Means
A centroid-based clustering algorithm that partitions
data into a predefined number of clusters by assigning
sample to the nearest center}
\usage{
Phenotype.Analysis.Kmeans(object)
}
\arguments{
\item{object}{A Ramanome object containing the dataset.}
}
\value{
The datafram which contains Kmeans result.
}
\description{
k-Means
A centroid-based clustering algorithm that partitions
data into a predefined number of clusters by assigning
sample to the nearest center
}
