% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotype-Analysis.R
\name{Phenotype.Analysis.Kmeans}
\alias{Phenotype.Analysis.Kmeans}
\title{k-Means Clustering Analysis}
\usage{
Phenotype.Analysis.Kmeans(object, k)
}
\arguments{
\item{object}{A Ramanome object containing the dataset.}

\item{k}{the number of clusters.}
}
\value{
The datafram which contains Kmeans result.
}
\description{
A centroid-based clustering algorithm that partitions
data into a predefined number of clusters by assigning
sample to the nearest center
}
\examples{
data(RamEx_data)
data_smoothed <- Preprocessing.Smooth.Sg(RamEx_data)
data_baseline <- Preprocessing.Baseline.Polyfit(data_smoothed)
data_baseline_bubble <- Preprocessing.Baseline.Bubble(data_smoothed)
data_normalized <- Preprocessing.Normalize(data_baseline, "ch")
data_cleaned <- Qualitycontrol.ICOD(data_normalized@datasets$normalized.data,var_tol = 0.4)
data_cleaned <- RamEx_data[qc_icod$quality,] 
data_cleaned <- Feature.Reduction.Intensity(data_cleaned, list(c(2000,2250),c(2750,3050), 1450, 1665))
clusters_kmneans <- Phenotype.Analysis.Kmeans(data_cleaned,5)
}
