% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotype-Analysis.R
\name{Phenotype.Analysis.Kmeans}
\alias{Phenotype.Analysis.Kmeans}
\title{k-Means Clustering Analysis}
\usage{
Phenotype.Analysis.Kmeans(object, n_pc = 10, k)
}
\arguments{
\item{object}{A Ramanome object.}

\item{n_pc}{The number of principal components to retain in the PCA. Defaults to 10.}

\item{k}{the number of clusters.}
}
\value{
A list containing:
\describe{ 
  \item{clusters}{The cluster assignments for each data point}
  \item{centers}{The coordinates of the cluster centers}
}
}
\description{
A centroid-based partitioning algorithm that assigns data points to clusters by minimizing the sum of squared distances between points and their cluster centers
}
\examples{
data(RamEx_data)
data_processed <- Preprocessing.OneStep(RamEx_data)
clusters_kmneans <- Phenotype.Analysis.Kmeans(data_processed,5)
}
