% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotype-Analysis.R
\name{Phenotype.Analysis.Louvaincluster}
\alias{Phenotype.Analysis.Louvaincluster}
\title{Louvain Clustering}
\usage{
Phenotype.Analysis.Louvaincluster(
  object,
  resolutions,
  n_pc = 10,
  threshold = 0.001,
  k = 30,
  n_tree = 50,
  seed = 42
)
}
\arguments{
\item{object}{A Ramanome object.}

\item{resolutions}{A vector of resolutions at which to perform the Louvain clustering.}

\item{n_pc}{The number of principal components to retain in the PCA. Defaults to 10.}

\item{threshold}{The minimum number of samples required to consider a cluster valid. Defaults to 0.001.}

\item{k}{Number of Nearest Neighbors, determines how many nearest neighbors each data point will consider when building the k-NN graph. A larger k will result in a denser graph, where each node is connected to more neighbors. This can capture broader local structures but may blur fine-grained distinctions. A smaller k will produce a sparser graph, emphasizing local structures but potentially missing broader patterns.}

\item{n_tree}{The number of trees to use in the Approximate Nearest Neighbor search. Defaults to 50.
A higher n_tree improves the accuracy of the nearest neighbor search but increases computation time and memory usage. A lower n_tree speeds up the search but may return less accurate neighbors.}

\item{seed}{The seed for the random number generator.}
}
\value{
A data frame containing the cluster assignments for each resolution.
}
\description{
A graph-based community detection algorithm that optimizes modularity to identify clusters by iteratively merging nodes to maximize within-cluster connections.
}
\examples{
data(RamEx_data)
data_processed <- Preprocessing.OneStep(RamEx_data)
#options(mc.cores = 2)
#clusters_Louvaincluster <- Phenotype.Analysis.Louvaincluster(object = data_processed, resolutions = c(0.8))
}
