% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irca.R
\name{Plot_local_chordDiagram_rpvalue}
\alias{Plot_local_chordDiagram_rpvalue}
\title{Plot local chord diagram with p-value thresholds}
\usage{
Plot_local_chordDiagram_rpvalue(
  x,
  Pos_Edge = FALSE,
  Neg_Edge = FALSE,
  Threshold = 0.6,
  bands_ann = NULL
)
}
\arguments{
\item{x}{A correlation matrix.}

\item{Pos_Edge}{Logical value to indicate whether to include positive edges. Default is FALSE.}

\item{Neg_Edge}{Logical value to indicate whether to include negative edges. Default is FALSE.}

\item{Threshold}{The threshold value for correlation. Default is 0.6.}

\item{bands_ann}{Annotation for the bands.}
}
\description{
This function plots a local chord diagram based on a correlation matrix and p-value thresholds.
}
