% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocesssing.R
\name{Preprocessing.Background.Remove}
\alias{Preprocessing.Background.Remove}
\title{Perform background subtraction on Ramanome object}
\usage{
Preprocessing.Background.Remove(object, cell.index, cal_mean = FALSE)
}
\arguments{
\item{object}{A Ramanome object.}

\item{cell.index}{The index of the cell component in the filenames.}

\item{cal_mean}{Logical value indicating whether to calculate the mean background spectrum.}
}
\value{
The modified Ramanome object.
}
\description{
This function performs background subtraction on a Ramanome object by subtracting the background spectra from the dataset.
}
