% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{Preprocessing.Baseline.Bubble}
\alias{Preprocessing.Baseline.Bubble}
\title{Remove baseline (low frequency noises) by bubble}
\usage{
Preprocessing.Baseline.Bubble(object, min_bubble_widths = 50)
}
\arguments{
\item{object}{Ramanome object}

\item{min_bubble_widths}{The minimum width of the bubbles.}
}
\value{
The Ramanome object with baseline corrected spectra, named as "baseline.data" in the 'datasets' slot
}
\description{
Bubble method is more suitable for spectra with complex background and uneven baseline. Bubble filling was performed on the 500–1800 cm⁻¹ region, while polynomial fitting was applied to the 1800–3050 cm⁻¹ region.
}
\examples{
data(RamEx_data)
data_baseline_bubble <- Preprocessing.Baseline.Bubble(RamEx_data)
}
