% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocesssing.R
\name{Preprocessing.Normalize}
\alias{Preprocessing.Normalize}
\title{Pre-process Raman data by ch normalization method}
\usage{
Preprocessing.Normalize(object, method, wave = NULL)
}
\arguments{
\item{object}{A Ramanome object.}

\item{method}{The normalize method}

\item{wave}{The interested wavenumber}
}
\value{
The updated Ramanome object with normalized Raman data.
}
\description{
This function applies normalization to the Raman data based on the ch method/max method/specific method/area method and updates the Raman object.
}
\examples{
data(RamEx_data)
data_smoothed <- Preprocessing.Smooth.Sg(RamEx_data)
data_baseline <- Preprocessing.Baseline.Polyfit(data_smoothed)
data_baseline_bubble <- Preprocessing.Baseline.Bubble(data_smoothed)
data_normalized_ch <- Preprocessing.Normalize(data_baseline, "ch")
#data_normalized_specific <- Preprocessing.Normalize(data_baseline, "specific")
data_normalized_max <- Preprocessing.Normalize(data_baseline, "max")
data_normalized_area <- Preprocessing.Normalize(data_baseline, "area")
}
