% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocesssing.R
\name{Preprocessing.Smooth.Sg}
\alias{Preprocessing.Smooth.Sg}
\title{Smooth Spectral Data Using Savitzky-Golay Filter}
\usage{
Preprocessing.Smooth.Sg(object, m = 0, p = 5, w = 11, delta.wav = 2)
}
\arguments{
\item{object}{A Ramanome object containing the spectral data.}

\item{m}{The order of the polynomial used in the Savitzky-Golay filter. Defaults to 0.}

\item{p}{The number of points to use in the Savitzky-Golay filter. Defaults to 5.}

\item{w}{The window size used in the Savitzky-Golay filter. Defaults to 11.}

\item{delta.wav}{The change in wavelength between each data point. Defaults to 2.}
}
\value{
The updated Ramanome object with the smoothed spectral data.
}
\description{
This function applies the Savitzky-Golay filter to the spectral data in a Ramanome
object to reduce noise. The filter is applied to the central portion of the data,
preserving the first and last few columns as specified by the `w` parameter.
}
\examples{
data(RamEx_data)
data_smoothed <- Preprocessing.Smooth.Sg(RamEx_data)
}
