% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{Preprocessing.Smooth.Snv}
\alias{Preprocessing.Smooth.Snv}
\title{Apply Standard Normal Variate (SNV) Transformation}
\usage{
Preprocessing.Smooth.Snv(object)
}
\arguments{
\item{object}{A Ramanome object.}
}
\value{
A Ramanome object with SNV-transformed spectra, named as "smooth.data" in the 'datasets' slot
}
\description{
This function applies the Standard Normal Variate (SNV) transformation to the raw
spectral data in a Ramanome object. The SNV transformation is a type of preprocessing
technique used to normalize the data and reduce the impact of multiplicative scatter.
The function also reports the number of spectra and the wavelength range after
the transformation.
}
\examples{
data(RamEx_data)
data_smoothed_snv <- Preprocessing.Smooth.Snv(RamEx_data)
}
