% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{Preprocessing.Spike}
\alias{Preprocessing.Spike}
\title{Pre spike the spec. with Rcpp}
\usage{
Preprocessing.Spike(object, device)
}
\arguments{
\item{object}{The rammanome class}

\item{device}{Parallel device, default is null for CPU, can be set to "GPU"}
}
\value{
A rammanome object
}
\description{
This function performs Pre spike function
}
\examples{
data(RamEx_data)
data_spike <- Preprocessing.Spike(RamEx_data,"CPU")
}
