% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocesssing.R
\name{Preprocesssing.Background.Spike}
\alias{Preprocesssing.Background.Spike}
\title{Pre spike the spec. with Rcpp}
\usage{
Preprocesssing.Background.Spike(object, device)
}
\arguments{
\item{object}{The rammanome class}

\item{device}{Parallel device, default is null for CPU, can be set to "GPU"}
}
\value{
A rammanome object
}
\description{
This function performs Pre spike function
}
