% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocesssing.R
\name{Preprocesssing.Cutoff}
\alias{Preprocesssing.Cutoff}
\title{Pre-process Raman data by cutting the wavenumber range}
\usage{
Preprocesssing.Cutoff(object, from, to)
}
\arguments{
\item{object}{A Ramanome object.}

\item{from}{The lower bound of the wavenumber range to cut.}

\item{to}{The upper bound of the wavenumber range to cut.}
}
\value{
The updated Ramanome object with the wavenumber range cut and the corresponding dataset.
}
\description{
This function cuts the wavenumber range of Raman data and updates the Raman object.
}
\examples{
data(RamEx_data)
data_smoothed <- Preprocesssing.Smooth.Sg(RamEx_data)
data_baseline <- Preprocesssing.Baseline.Polyfit(data_smoothed)
data_baseline_bubble <- Preprocesssing.Baseline.Bubble(data_smoothed)
data_normalized <- Preprocesssing.Normalize(data_baseline, "ch")
Preprocesssing.Cutoff(data_normalized,550, 1800)
}
