% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocesssing.R
\name{Preprocesssing.Normalize}
\alias{Preprocesssing.Normalize}
\title{Pre-process Raman data by ch normalization method}
\usage{
Preprocesssing.Normalize(object, method, wave = NULL)
}
\arguments{
\item{object}{A Ramanome object.}

\item{method}{The normalize method}

\item{wave}{The interested wavenumber}
}
\value{
The updated Ramanome object with normalized Raman data.
}
\description{
This function applies normalization to the Raman data based on the ch method/max method/specific method/area method and updates the Raman object.
}
