% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality-Control.R
\name{Qualitycontrol.Mcd}
\alias{Qualitycontrol.Mcd}
\title{Detect outliers using the Minimum Covariance Determinant (MCD) method}
\usage{
Qualitycontrol.Mcd(x, index_good, h = 0.5, alpha = 0.01, na.rm = TRUE)
}
\arguments{
\item{x}{The input dataset.}

\item{index_good}{A vector indicating the indices of good observations.}

\item{h}{The fraction of the data to use for computing the MCD.}

\item{alpha}{The significance level for the MCD.}

\item{na.rm}{A logical value indicating whether to remove NA values.}
}
\value{
An object of class "Qualitycontrol.Mcd" containing the following components:
  - distance: The maximum distance threshold for outliers.
  - center: The center of the outliers.
  - outliers_pos: The positions of the outliers.
  - nb: A named number giving the total number of outliers detected.
}
\description{
This function extends the outliers_mcdEst function to detect outliers in a dataset
using the Minimum Covariance Determinant method. It modifies the output to include
additional information such as the distance and center of the outliers.
}
\examples{
# Create sample multivariate data
set.seed(123)
n <- 100
p <- 5
X <- matrix(rnorm(n*p), ncol=p)

# Add some outliers
X[1:5,] <- X[1:5,] + 10

# Create index of initially good observations
index_good <- rep(TRUE, n)

}
