% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality-Control.R
\name{Qualitycontrol.Snr}
\alias{Qualitycontrol.Snr}
\title{Calculate the Signal-to-Noise Ratio (SNR) for a dataset}
\usage{
Qualitycontrol.Snr(object, level = "medium")
}
\arguments{
\item{object}{Ramanome Object}

\item{level}{The level of strictness for the SNR calculation. Defaults to "medium".}
}
\value{
A data frame with quality defined by SNR method and their SNR values.
}
\description{
This function calculates the SNR for a dataset by comparing the signal and noise levels.
The signal is defined as the maximum intensity in the range of 600-3050 cm-1, and the noise is defined as the mean intensity in the range of 1800-1900 cm-1.
The SNR is calculated as the ratio of the mean of signal and the standard deviation of noise.
}
\examples{
data(RamEx_data)
qc_snr <- Qualitycontrol.Snr(RamEx_data)
}
