% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantification.R
\name{Quantification.Glm}
\alias{Quantification.Glm}
\title{Generalized linear model (GLM)
Extends linear regression by allowing response variables to follow non-normal distributions via link functions.}
\usage{
Quantification.Glm(train, test = NULL, n_pc = 20, seed = 42)
}
\arguments{
\item{train}{The training data object}

\item{test}{The test data object (optional). If not provided, the function will perform a stratified cross-validation (Training : Test = 7 : 3).}

\item{n_pc}{The number of principal components that the PCA model extract for further GLM model building. Feature extraction is recommended to reduce the multicollinearity among predictors.}

\item{seed}{The seed for the random number generator}
}
\value{
A list containing:
\describe{ 
  \item{model}{The GLM model}
  \item{pred_test}{The prediction for test data if test is provided}
  \item{pca_params}{The pre-feature extraction PCA parameters used for the GLM model building}
}
}
\description{
Generalized linear model (GLM)
Extends linear regression by allowing response variables to follow non-normal distributions via link functions.
}
\examples{
data(RamEx_data)
data_processed <- Preprocessing.OneStep(RamEx_data)
quan_glm <- Quantification.Glm(data_processed)
}
