% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantification.R
\name{Quantification.Glm}
\alias{Quantification.Glm}
\title{Generalized linear model (GLM)
SExtends linear regression by allowing the dependent
variable to follow distributions other than
normal (e.g., binomial, Poisson) and uses a link function
 to relate predictors to the response}
\usage{
Quantification.Glm(train, test = NULL)
}
\arguments{
\item{train}{The training data object}

\item{test}{The test data object (optional)}
}
\value{
The GLM prediction result.
}
\description{
Generalized linear model (GLM)
SExtends linear regression by allowing the dependent
variable to follow distributions other than
normal (e.g., binomial, Poisson) and uses a link function
 to relate predictors to the response
}
\examples{
data(RamEx_data)
data_smoothed <- Preprocessing.Smooth.Sg(RamEx_data)
data_baseline <- Preprocessing.Baseline.Polyfit(data_smoothed)
data_baseline_bubble <- Preprocessing.Baseline.Bubble(data_smoothed)
data_normalized <- Preprocessing.Normalize(data_baseline, "ch")
data_cleaned <- Qualitycontrol.ICOD(data_normalized@datasets$normalized.data,var_tol = 0.4)
data_cleaned <- RamEx_data[qc_icod$quality,]
data_cleaned <- Feature.Reduction.Intensity(data_cleaned, list(c(2000,2250),c(2750,3050), 1450, 1665))
data_cleaned <- Feature.Reduction.Pca(data_cleaned, draw=FALSE , save = FALSE)
quan_glm <- Quantification.Glm(data_cleaned)
}
