% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantification.R
\name{Quantification.Mlr}
\alias{Quantification.Mlr}
\title{Multiple linear regression (MLR)
Simple and interpretable regression method, but assumes
no multicollinearity among predictors and a linear
relationship between predictors and response}
\usage{
Quantification.Mlr(train, test = NULL)
}
\arguments{
\item{train}{The training data object}

\item{test}{The test data object (optional)}
}
\value{
The MLR prediction result.
}
\description{
Multiple linear regression (MLR)
Simple and interpretable regression method, but assumes
no multicollinearity among predictors and a linear
relationship between predictors and response
}
\examples{
data(RamEx_data)
data_smoothed <- Preprocessing.Smooth.Sg(RamEx_data)
data_baseline <- Preprocessing.Baseline.Polyfit(data_smoothed)
data_baseline_bubble <- Preprocessing.Baseline.Bubble(data_smoothed)
data_normalized <- Preprocessing.Normalize(data_baseline, "ch")
data_cleaned <- Qualitycontrol.ICOD(data_normalized@datasets$normalized.data,var_tol = 0.4)
data_cleaned <- RamEx_data[qc_icod$quality,]
data_cleaned <- Feature.Reduction.Intensity(data_cleaned, list(c(2000,2250),c(2750,3050), 1450, 1665))
data_cleaned <- Feature.Reduction.Pca(data_cleaned,  draw=FALSE , save = FALSE)
quan_mlr <- Quantification.Mlr(data_cleaned)
}
