% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantification.R
\name{Quantification.Mlr}
\alias{Quantification.Mlr}
\title{Multiple linear regression (MLR)
Simple and interpretable regression method, but assumes
no multicollinearity among predictors and a linear
relationship between predictors and response}
\usage{
Quantification.Mlr(train, test = NULL)
}
\arguments{
\item{train}{The training data object}

\item{test}{The test data object (optional)}
}
\value{
The MLR prediction result.
}
\description{
Multiple linear regression (MLR)
Simple and interpretable regression method, but assumes
no multicollinearity among predictors and a linear
relationship between predictors and response
}
