% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantification.R
\name{Quantification.Pls}
\alias{Quantification.Pls}
\title{Partial Least Squares (PLS)}
\usage{
Quantification.Pls(train, test = NULL, n_comp = 8, seed = 42)
}
\arguments{
\item{train}{The training Ramanome object}

\item{test}{The test data object (optional). If not provided, the function will perform a stratified cross-validation (Training : Test = 7 : 3).}

\item{n_comp}{The number of latent components that the PLS model extract from the training data to model the response variable}

\item{seed}{The seed for the random number generator}
}
\value{
A list containing:
\describe{ 
  \item{model}{The PLS model}
  \item{pred_test}{The prediction for test data if test is provided}
}
}
\description{
A regression method that projects predictors and responses to latent variables to maximize covariance.
}
\examples{
data(RamEx_data)
data_processed <- Preprocessing.OneStep(RamEx_data)
quan_pls <- Quantification.Pls(data_processed)
}
