% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantification.R
\name{Quantification.Pls}
\alias{Quantification.Pls}
\title{Partial Least Squares (PLS)}
\usage{
Quantification.Pls(train, test = NULL)
}
\arguments{
\item{train}{The training data object}

\item{test}{The test data object (optional)}
}
\value{
The PLS model.
}
\description{
A regression method that reduces predictors to a smaller set of latent
variables while maximizing the covariance between predictors and response
 variables, ideal for situations where predictors are highly collinear
 or when the number of predictors exceeds the number of observations
}
