% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raman-Markers.R
\name{Raman.Markers.Correlations}
\alias{Raman.Markers.Correlations}
\title{Compute Correlations with a Target Variable}
\usage{
Raman.Markers.Correlations(object, min.cor = 0.8)
}
\arguments{
\item{object}{A Ramanome object.}

\item{min.cor}{The minimum correlation threshold. Defaults to 0.8.}
}
\value{
A list containing the high correlation elements and their corresponding correlations with the target variable, containing two data frames: 'correlations_singular' for single markers (group ~ wave) and 'combination_correlations_two' for paired markers (group ~ wave1 / wave2).
}
\description{
Calculates the correlations between each column of a dataset X and a target variable y by traversing all possible single and paired markers.
}
\examples{
data(RamEx_data)
data_processed <- Preprocessing.OneStep(RamEx_data)
options(mc.cores = 2)
cor_markers <- Raman.Markers.Correlations(preprocessing.cutoff(data_processed, 800, 1300), min.cor = 0.6)
}
