% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raman-Markers.R
\name{Raman.Markers.Rbcs}
\alias{Raman.Markers.Rbcs}
\title{Perform random forest classification and feature importance analysis}
\usage{
Raman.Markers.Rbcs(
  object,
  outfolder = "RF_imps",
  draw = TRUE,
  threshold = 0.002,
  ntree = 1000,
  errortype = "oob",
  verbose = FALSE,
  nfolds = 3,
  rf.cv = FALSE
)
}
\arguments{
\item{object}{A Ramanome object.}

\item{outfolder}{The output folder to save the results.}

\item{draw}{Logical value indicating whether to draw a heatmap.}

\item{threshold}{The threshold value for feature importances.}

\item{ntree}{The number of trees in the random forest model.}

\item{errortype}{The error type used in random forest model evaluation.}

\item{verbose}{Logical value indicating whether to display progress messages.}

\item{nfolds}{The number of folds used in cross-validation.}

\item{rf.cv}{Logical value indicating whether to perform random forest with cross-validation.}
}
\value{
A data frame containing the selected feature importances.
}
\description{
This function performs random forest classification on Raman data and generates feature importances.
}
\examples{
data(RamEx_data)
data_smoothed <- Preprocessing.Smooth.Sg(RamEx_data)
data_baseline <- Preprocessing.Baseline.Polyfit(data_smoothed)
data_baseline_bubble <- Preprocessing.Baseline.Bubble(data_smoothed)
data_normalized <- Preprocessing.Normalize(data_baseline, "ch")
data_cleaned <- Qualitycontrol.ICOD(data_normalized@datasets$normalized.data,var_tol = 0.4)
data_cleaned <- RamEx_data[qc_icod$quality,]
data_cleaned <- Feature.Reduction.Intensity(data_cleaned, list(c(2000,2250),c(2750,3050), 1450, 1665))
RBCS.markers <- Raman.Markers.Rbcs(data_cleaned, threshold = 0.003, draw = FALSE)
}
