% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamEx-Class.R
\docType{class}
\name{Ramanome-class}
\alias{Ramanome-class}
\alias{Ramanome}
\title{Ramanome class}
\value{
A new Ramanome object
}
\description{
This class represents a Ramanome object, which contains Raman data and associated metadata.
}
\section{Slots}{

\describe{
\item{\code{wavenumber}}{Numeric vector indicating the wavenumbers of the Raman spectra}

\item{\code{datasets}}{A list of Raman datasets}

\item{\code{meta.data}}{A data frame containing metadata associated with the Raman datasets}

\item{\code{reductions}}{A list of data reduction results}

\item{\code{interested.bands}}{A list of interested Raman bands}
}}

\examples{
# Create sample data
wavenumbers <- seq(500, 3500, by = 10)
spectra <- matrix(rnorm(100 * length(wavenumbers)), nrow = 100)
metadata <- data.frame(
  group = factor(rep(c("Control", "Treatment"), each = 50)),
  filenames = paste0("sample_", 1:100, ".txt")
)

# Create a Ramanome object
raman_obj <- new("Ramanome",
  datasets = list(raw = spectra),
  wavenumber = wavenumbers,
  meta.data = metadata
)

}
