% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rammanome-class.R
\docType{class}
\name{Ramanome-class}
\alias{Ramanome-class}
\alias{Ramanome}
\title{Ramanome class}
\description{
This class represents a Ramanome object, which contains Raman data and associated metadata.
}
\section{Slots}{

\describe{
\item{\code{wavenumber}}{Numeric vector indicating the wavenumbers of the Raman spectra}

\item{\code{datasets}}{A list of Raman datasets}

\item{\code{meta.data}}{A data frame containing metadata associated with the Raman datasets}

\item{\code{reductions}}{A list of data reduction results}

\item{\code{interested.bands}}{A list of interested Raman bands}
}}

