% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rammanome-class.R
\name{rbind2,Ramanome,Ramanome-method}
\alias{rbind2,Ramanome,Ramanome-method}
\title{#' Plot method for Ramanome objects
#'
#' @param x The Ramanome object
#' @param y Character specifying the plot type
#' @importFrom hyperSpec plot
#' @importFrom hyperSpec print
#' @importFrom methods setMethod}
\usage{
\S4method{rbind2}{Ramanome,Ramanome}(x, y)
}
\arguments{
\item{x}{The first Ramanome object}

\item{y}{The second Ramanome object}
}
\description{
setMethod("plot",
          signature(x = "Ramanome", y = "character"),
          function(x, y, ...) {
            tmp <- hyperSpec::plot(x, y, ...)
            if (is(tmp, "trellis"))
              hyperSpec::print(tmp)
            invisible(tmp)
          }
)
Combine Ramanome objects by row binding
}
