% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-spec.R
\name{read.single}
\alias{read.single}
\title{Read and Process a Single Spectral File}
\usage{
read.single(cell_path, xout)
}
\arguments{
\item{cell_path}{The file path to the spectral data file. If the path is invalid or empty,
the function will return an empty result.}

\item{xout}{A vector of wavelengths at which to estimate the intensity values.}
}
\value{
a dataframe
}
\description{
This function reads a spectral file from a specified path, checks if the path is valid,
and then uses spline interpolation to create a function that can be used to estimate
intensity values at a specified set of wavelengths (xout).
}
