% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamEx-Basic-Funciton.R
\name{save.renishaw}
\alias{save.renishaw}
\title{Save Renishaw Raman data}
\usage{
save.renishaw(dir_path)
}
\arguments{
\item{dir_path}{The directory path where the Raman data is located.}
}
\value{
NULL. The function processes the input text files and saves new text files in the same directory
        with modified names based on the grouping of the data. The original input files are removed.
}
\description{
This function saves the Renishaw Raman data after processing.
}
