% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raman-Markers.R
\name{save.rf.results.confusion.matrix}
\alias{save.rf.results.confusion.matrix}
\title{Save confusion matrix of random forest classification results}
\usage{
save.rf.results.confusion.matrix(
  result,
  filename = "confusion_matrix.xls",
  outdir
)
}
\arguments{
\item{result}{The result object obtained from random forest classification}

\item{filename}{The name of the output file}

\item{outdir}{The path to the output directory}
}
\value{
None
}
\description{
This function saves the confusion matrix of random forest classification results to a file.
}
\examples{
# Create sample results with confusion matrix
result <- list(
  confusion.matrix = matrix(c(45, 5, 8, 42), 2, 2,
                           dimnames=list(c("Class1", "Class2"),
                                        c("Class1", "Class2")))
)

}
