% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raman-Markers.R
\name{save.rf.results.mislabeling}
\alias{save.rf.results.mislabeling}
\title{Save mislabeling scores from random forest classification}
\usage{
save.rf.results.mislabeling(result, filename = "mislabeling.xls", outdir)
}
\arguments{
\item{result}{The result object from random forest classification}

\item{filename}{The filename for the probabilities file (default is "mislabeling.xls")}

\item{outdir}{The output directory to save the probabilities file}
}
\value{
NULL. The function saves a tab-delimited text file containing mislabeling scores
        for each sample, with columns for sample ID and corresponding mislabeling score.
}
\description{
Save mislabeling scores from random forest classification
}
