% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raman-Markers.R
\name{save.rf.results.mislabeling}
\alias{save.rf.results.mislabeling}
\title{Title}
\usage{
save.rf.results.mislabeling(result, filename = "mislabeling.xls", outdir)
}
\arguments{
\item{result}{The result object from random forest classification}

\item{filename}{The filename for the probabilities file (default is "mislabeling.xls")}

\item{outdir}{The output directory to save the probabilities file}
}
\description{
Title
}
