% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raman-Markers.R
\name{save.rf.results.probabilities}
\alias{save.rf.results.probabilities}
\title{Save probabilities of random forest classification results}
\usage{
save.rf.results.probabilities(
  result,
  filename = "cv_probabilities.xls",
  outdir
)
}
\arguments{
\item{result}{The result object from random forest classification}

\item{filename}{The filename for the probabilities file (default is "cv_probabilities.xls")}

\item{outdir}{The output directory to save the probabilities file}
}
\value{
NULL. The function saves a tab-delimited text file containing class probabilities
        for each sample, with rows for samples and columns for class probabilities.
}
\description{
This function saves the probabilities of random forest classification results to a file.
}
\examples{
# Create sample results with probabilities
result <- list(
  probabilities = matrix(runif(100), 10, 10,
                        dimnames=list(paste0("Sample", 1:10),
                                     paste0("Class", 1:10)))
)

}
