% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamEx-Basic-Funciton.R
\name{select.value}
\alias{select.value}
\title{Select a single Raman Feature.Reduction.Intensity value for a given wavenumber}
\usage{
\method{select}{value}(object, wave)
}
\arguments{
\item{object}{A Ramanome object.}

\item{wave}{The wavenumber for which to retrieve the Feature.Reduction.Intensity value.}
}
\value{
The Raman Feature.Reduction.Intensity value for the given wavenumber.
}
\description{
This function selects a single Raman Feature.Reduction.Intensity value for a given wavenumber from a Ramanome object.
}
