% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_value.R
\name{select.value}
\alias{select.value}
\title{Select a single Raman intensity value for a given wavenumber}
\usage{
\method{select}{value}(object, wave)
}
\arguments{
\item{object}{A Ramanome object.}

\item{wave}{The wavenumber for which to retrieve the intensity value.}
}
\value{
The Raman intensity value for the given wavenumber.
}
\description{
This function selects a single Raman intensity value for a given wavenumber from a Ramanome object.
}
