% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-new-class.R, R/base-ramanome-class.R
\name{show,Ramanome-method}
\alias{show,Ramanome-method}
\title{Show method for Ramanome class}
\usage{
\S4method{show}{Ramanome}(object)

\S4method{show}{Ramanome}(object)
}
\arguments{
\item{object}{The Ramanome object}
}
\description{
This method shows information about the Ramanome object.
}
