% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamEx-Class.R
\name{show,Ramanome-method}
\alias{show,Ramanome-method}
\title{Show method for Ramanome objects
This method will give brief information about the Ramanome object}
\usage{
\S4method{show}{Ramanome}(object)
}
\arguments{
\item{object}{The Ramanome object}
}
\value{
No return value, prints object information to console
}
\description{
Show method for Ramanome objects
This method will give brief information about the Ramanome object
}
\examples{
# Create sample Ramanome object
wavenumbers <- seq(500, 3500, by = 10)
spectra <- matrix(rnorm(100 * length(wavenumbers)), nrow = 100)
metadata <- data.frame(
  group = factor(rep(c("Control", "Treatment"), each = 50))
)
raman_obj <- new("Ramanome",
  datasets = list(raw = spectra),
  wavenumber = wavenumbers,
  meta.data = metadata
)

# Display object information
show(raman_obj)
}
