% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rammanome-class.R
\name{show,Ramanome-method}
\alias{show,Ramanome-method}
\title{Show method for Ramanome objects
This method will give brief information about the Ramanome object}
\usage{
\S4method{show}{Ramanome}(object)
}
\arguments{
\item{object}{The Ramanome object}
}
\description{
Show method for Ramanome objects
This method will give brief information about the Ramanome object
}
