% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamEx-Basic-Funciton.R
\name{spec.mean.draw}
\alias{spec.mean.draw}
\title{Plot and Save the Mean Spectrum}
\usage{
spec.mean.draw(object, gap = 0)
}
\arguments{
\item{object}{A Ramanome object containing the spectral data and metadata.}

\item{gap}{An optional numeric value specifying the gap between groups in the plot.
Defaults to 0.}
}
\description{
This function generates a plot of the mean spectrum for each group in the dataset
contained within a Ramanome object. It uses the `mean.spec` function to calculate
the mean spectrum and then saves the plot as a PNG file.
}
