% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectral-Decomposition.R
\name{Spectral.Decomposition.Ica}
\alias{Spectral.Decomposition.Ica}
\title{Perform Independent Component Analysis (ICA)}
\usage{
Spectral.Decomposition.Ica(object, ncomp)
}
\arguments{
\item{object}{A Ramanome object containing the dataset.}

\item{ncomp}{The number of components to retain in the ICA decomposition.}
}
\value{
A list containing the results of the ICA decomposition.
}
\description{
This function applies the FastICA algorithm to the dataset retrieved from the Ramanome
object using `get.nearest.dataset`. It returns the result of the ICA decomposition.
}
