% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectral-Decomposition.R
\name{Spectral.Decomposition.Ica}
\alias{Spectral.Decomposition.Ica}
\title{Independent Component Analysis (ICA)}
\usage{
Spectral.Decomposition.Ica(object, n_comp = 2, seed = 42)
}
\arguments{
\item{object}{A Ramanome object.}

\item{n_comp}{The number of components to retain in the ICA decomposition.}

\item{seed}{The seed for the random number generator.}
}
\value{
A list containing the results of the ICA decomposition.
}
\description{
A statistical technique that separates a multivariate signal into additive, statistically independent components, assuming non-Gaussian sources.
}
\examples{
data(RamEx_data)
data_processed <- Preprocessing.OneStep(RamEx_data)
decom_ica <- Spectral.Decomposition.Ica(data_processed, 2)
}
