% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectral-Decomposition.R
\name{Spectral.Decomposition.Mcrals}
\alias{Spectral.Decomposition.Mcrals}
\title{Perform Multivariate Curve Resolution Alternating Least Squares (MCR-ALS)}
\usage{
Spectral.Decomposition.Mcrals(object, ncomp)
}
\arguments{
\item{object}{A Ramanome object containing the dataset.}

\item{ncomp}{The number of components to estimate.}
}
\value{
A list containing the estimated components and concentrations.
}
\description{
This function applies the MCR-ALS algorithm to the dataset stored in a Ramanome object.
MCR-ALS is a chemometric technique used to resolve overlapping spectra into their
constituent pure components. The function returns the estimated components and
concentrations.
}
\examples{
data(RamEx_data)
data_processed <- Preprocessing.OneStep(RamEx_data)
decom_mcrals <- Spectral.Decomposition.Mcrals(data_processed, 2)
}
