% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectral-Decomposition.R
\name{Spectral.Decomposition.Mcrals}
\alias{Spectral.Decomposition.Mcrals}
\title{Multivariate Curve Resolution Alternating Least Squares (MCR-ALS)}
\usage{
Spectral.Decomposition.Mcrals(object, n_comp = 2, seed = 42)
}
\arguments{
\item{object}{A Ramanome object.}

\item{n_comp}{The number of components to estimate.}

\item{seed}{The seed for the random number generator.}
}
\value{
A list containing: 
\item{components}{The estimated components (n_wavelength x n_comp).} 
\item{concentration}{The estimated concentrations of each cell (n_cell x n_comp).}
while the reconstructed matrix is the product of the concentration and component matrices.
}
\description{
A chemometric method that decomposes a overlapping spectra matrix into concentration and pure spectra profiles iteratively under constraints like non-negativity.
This function applies the MCR-ALS algorithm to the dataset stored in a Ramanome object.
}
\examples{
data(RamEx_data)
data_processed <- Preprocessing.OneStep(RamEx_data)
decom_mcrals <- Spectral.Decomposition.Mcrals(data_processed, 2)
}
