% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectral-Decomposition.R
\name{Spectral.Decomposition.Nmf}
\alias{Spectral.Decomposition.Nmf}
\title{Perform Non-negative Matrix Factorization (NMF)}
\usage{
Spectral.Decomposition.Nmf(object)
}
\arguments{
\item{object}{A Ramanome object containing the dataset.}
}
\value{
A list containing the results of the NMF decomposition, including the basis
matrices and the reconstructed matrix.
}
\description{
This function applies Non-negative Matrix Factorization to the dataset retrieved from
a Ramanome object. NMF is a technique used to decompose a non-negative matrix into
two non-negative matrices, often used for data dimensionality reduction and feature
extraction.
}
\examples{
data(RamEx_data)
data_smoothed <- Preprocesssing.Smooth.Sg(RamEx_data)
data_baseline <- Preprocesssing.Baseline.Polyfit(data_smoothed)
data_baseline_bubble <- Preprocesssing.Baseline.Bubble(data_smoothed)
data_normalized <- Preprocesssing.Normalize(data_baseline, "ch")
data_cleaned <- Qualitycontrol.ICOD(data_normalized@datasets$normalized.data,var_tol = 0.4)
data_cleaned <- data_normalized[data_cleaned$index_good,]
data_cleaned <- Feature.Reduction.Intensity(data_cleaned, list(c(2000,2250),c(2750,3050), 1450, 1665))
decom_nmf <- Spectral.Decomposition.Nmf(data_cleaned)
}
