% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectral-Decomposition.R
\name{Spectral.Decomposition.Nmf}
\alias{Spectral.Decomposition.Nmf}
\title{Perform Non-negative Matrix Factorization (NMF)}
\usage{
Spectral.Decomposition.Nmf(object)
}
\arguments{
\item{object}{A Ramanome object containing the dataset.}
}
\value{
A list containing the results of the NMF decomposition, including the basis
matrices and the reconstructed matrix.
}
\description{
This function applies Non-negative Matrix Factorization to the dataset retrieved from
a Ramanome object. NMF is a technique used to decompose a non-negative matrix into
two non-negative matrices, often used for data dimensionality reduction and feature
extraction.
}
\examples{
data(RamEx_data)
data_processed <- Preprocessing.OneStep(RamEx_data)
decom_nmf <- Spectral.Decomposition.Nmf(data_processed)
}
