% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamEx-Class.R
\name{[,Ramanome,ANY,ANY,ANY-method}
\alias{[,Ramanome,ANY,ANY,ANY-method}
\title{Subset method for Ramanome objects}
\usage{
\S4method{[}{Ramanome,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{The Ramanome object}

\item{i}{The indices for subsetting rows}

\item{j}{The indices for subsetting columns (wavenumbers)}

\item{...}{Additional arguments}

\item{drop}{Boolean indicating whether to drop unused slots (default: TRUE)}
}
\value{
A new Ramanome object containing the subset of data
}
\description{
Subset method for Ramanome objects
}
