% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamEx-Basic-Funciton.R
\name{time.series}
\alias{time.series}
\title{Generate time series plots for Raman data}
\usage{
\method{time}{series}(object, reduction = "UMAP")
}
\arguments{
\item{object}{A Ramanome object.}

\item{reduction}{The reduction method to use (default is "UMAP").}
}
\value{
A cowplot object containing multiple plots arranged in a grid, with each plot showing the time series
        visualization for a different group in the dataset. The function also saves the plot as a PNG file
        named 'Time.series_[reduction].png'.
}
\description{
This function generates time series plots for Raman data using a specified reduction method (e.g., UMAP).
}
