% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intra-Ramanome-Analysis.R
\name{vectorize_dm_rcorr}
\alias{vectorize_dm_rcorr}
\title{Vectorize Correlation Coefficients and p-values from a Correlation Matrix}
\usage{
vectorize_dm_rcorr(dm, group = NULL, duplicate = TRUE)
}
\arguments{
\item{dm}{A list containing correlation coefficients (r) and p-values (P).}

\item{group}{A vector specifying the sample categories. (default is NULL)}

\item{duplicate}{A logical value indicating whether to include duplicate pairs. (default is TRUE)}
}
\value{
A data frame containing the vectorized correlation coefficients and p-values.
}
\description{
This function vectorizes the correlation coefficients and p-values from a correlation matrix,
optionally grouping the samples by a metadata category.
}
